### 2025-09-25 #9079 @editorialbot
[PRE REVIEW]: qlass: A Python Package for Variational Quantum Algorithms on Photonic Devices
**Submitting author:** @FarLab (<a href="http://orcid.org/0009-0005-3729-9602">Farrokh Labib</a>)
**Repository:** https://github.com/unitaryfoundation/qlass
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
