### 2025-09-26 ecc9aa1
Update paper.md
Fix missing references
