### 2025-09-26 039808b
Update paper.md
remove the folder structure
