### 2025-09-26 33212e0
Update paper.md
Made some edits. Added two sections to provide some more context in line with some guidelines by JOSS that advice to provide evidence for usage adoption.
