### 2025-09-26 e00f5c1
Update paper.md
small edit to fix enumeration on line 123 of paper.md
