### 2025-09-29 e589717
Bump perceval-quandela from 0.13.2 to 1.0.1
Bumps [perceval-quandela](https://github.com/Quandela/Perceval) from 0.13.2 to 1.0.1.
- [Release notes](https://github.com/Quandela/Perceval/releases)
- [Commits](https://github.com/Quandela/Perceval/compare/v0.13.2...v1.0.1)

---
updated-dependencies:
- dependency-name: perceval-quandela
  dependency-version: 1.0.1
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
