+++ docs/requirements.txt
@@ -1,7 +1,7 @@
 sphinx_rtd_theme
 qiskit==2.1.2
 numpy
-perceval-quandela==0.13.2
+perceval-quandela==1.0.1
 openfermion
 openfermionpyscf
 pyscf
