### 2025-10-01 0a2f62c
docs: add PyPI badge and stable install instructions to README and docs/installation.rst
