### 2025-10-02 fdc7075
add compute_expectation_value_from_unitary and loss_function_matrix to utils.py
