### 2025-10-06 bc8a070
enhance VQE class to work with unitary executors (instead of just samplers). add example file to showcase.
