### 2025-10-08 842491f
update noisy_example.py
