+++ examples/noisy_example.py
@@ -1,5 +1,7 @@
 """
-Example demonstrating the use of the VQE class with a noisy ansatz.
+This script performs a noisy VQE simulation for the tapered LiH Hamiltonian
+at various internuclear distances using the qlass library. It compares the
+simulated energy curve with the theoretical one.
 """
 
 import warnings
@@ -8,86 +10,114 @@ warnings.filterwarnings('ignore')
 
 import numpy as np
 import matplotlib.pyplot as plt
+from tqdm import tqdm
+from perceval import LogicalState
 from perceval.algorithm import Sampler
-from perceval.utils import NoiseModel
+from qiskit import transpile
+from qiskit.circuit.library import TwoLocal
 
-from qlass.quantum_chemistry import LiH_hamiltonian, brute_force_minimize
-from qlass.vqe import VQE, le_ansatz
+# Imports from the qlass library
+from qlass.quantum_chemistry import LiH_hamiltonian_tapered, brute_force_minimize
+from qlass.vqe import VQE
+from qlass.compiler import HardwareConfig, ResourceAwareCompiler
+from qlass.utils import rotate_qubits
 
-def executor_with_noise(params, pauli_string, noise_model):
+# Define an executor function that uses the TwoLocal ansatz
+def executor(params, pauli_string):
     """
-    Executor function that creates a processor with the le_ansatz,
-    applies the specified noise model, and runs the simulation.
+    Executor function that creates a TwoLocal ansatz and compiles it
+    using the ResourceAwareCompiler for resource analysis.
     """
-    processor = le_ansatz(params, pauli_string, noise_model=noise_model)
+    # Create the TwoLocal ansatz circuit
+    num_qubits = len(pauli_string)
+    ansatz = TwoLocal(num_qubits, 'ry', 'cx', reps=1)
+    
+    # Assign parameters to the ansatz
+    ansatz_assigned = ansatz.assign_parameters(params)
+    ansatz_transpiled = transpile(ansatz_assigned, basis_gates=['u3', 'cx'], optimization_level=3)
+    
+    # Apply rotation for Pauli measurement
+    ansatz_rot = rotate_qubits(pauli_string, ansatz_transpiled.copy())
+    
+    # Define hardware configuration for the example photonic chip
+    chip_config = HardwareConfig(
+        brightness=0.09,
+        indistinguishability=0.9,
+        g2=0.01,
+        g2_distinguishable=True,
+        transmittance=0.4,
+        phase_imprecision=0.02,
+        phase_error=0.02
+    )
+    
+    # Compile with ResourceAwareCompiler
+    compiler = ResourceAwareCompiler(config=chip_config)
+    processor = compiler.compile(ansatz_rot)
+
+    # Set the input state
+    processor.with_input(LogicalState([0]*num_qubits))
+    
+    # Run the sampler
     sampler = Sampler(processor)
     samples = sampler.samples(10_000)
     return samples
 
 def main():
-    # --- 1. Define the Problem ---
-    num_qubits = 2
-    hamiltonian = LiH_hamiltonian(num_electrons=2, num_orbitals=1)
+    """
+    Main function to run the VQE simulation and plot the results.
+    """
     
-    print("LiH Hamiltonian:")
-    for pauli_string, coefficient in hamiltonian.items():
-        print(f"  {pauli_string}: {coefficient:.4f}")
+    # 1. Define the range of bond radii to simulate
+    radii = np.linspace(0.5, 2.5, 15)
+    exact_energies = []
+    vqe_energies = []
     
-    exact_energy = brute_force_minimize(hamiltonian)
-    print(f"\nExact ground state energy: {exact_energy:.6f}")
-    
[…]
