### 2025-10-08 35b3f79
remove an example file
