### 2025-10-13 aa61cf7
Added HF initial states, KS Hamiltonian, Ensemble VQE functions and example files
