+++ qlass/vqe/ansatz.py
@@ -1,13 +1,14 @@
 from perceval.converters import QiskitConverter
 from perceval.utils import NoiseModel
 from qiskit import transpile, QuantumCircuit
-from qiskit.circuit.library import n_local
-
+from qiskit.circuit.library import TwoLocal
+from qiskit.quantum_info import Statevector
 import perceval as pcvl
 import numpy as np
 from perceval.components import Processor
 from qlass.utils import rotate_qubits
 from qlass.compiler import compile
+
 
 def le_ansatz(lp: np.ndarray, pauli_string: str, noise_model: NoiseModel = None) -> Processor:
     """
@@ -71,3 +72,80 @@ def custom_unitary_ansatz(lp: np.ndarray,
     processor.with_input(pcvl.LogicalState([0]*num_qubits))
 
     return processor
+
+
+def list_of_ones(computational_basis_state: int, n_qubits):
+    """
+    Indices of ones in the binary expansion of an integer in big endian
+    order. e.g. 010110 -> [1, 3, 4] (which is the reverse of the qubit ordering...)
+    """
+
+    bitstring = format(computational_basis_state, 'b').zfill(n_qubits)
+
+    return [abs(j - n_qubits + 1) for j in range(len(bitstring)) if bitstring[j] == '1']
+
+
+
+def hf_ansatz(layers: int, n_orbs: int, lp: np.ndarray, pauli_string: str, method: str, noise_model: NoiseModel = None) -> Processor:
+    """
+    Args:
+        layers (int): number of circuit layers
+        n_orbs (int): number of orbitals
+        lp (np.ndarray): array of parameter values
+        pauli_string (str): Pauli string
+        method (str): 'WFT' n_occ spin orbitals mapped into n_aubits, 'DFT' spatial orbitals mapped into log2(N) qubits by considering only spin-alpha or spin-beta blocks
+        see article https://scipost.org/SciPostPhys.14.3.055 for DFT mapping.
+        noise_model (NoiseModel): A perceval NoiseModel object representing the noise model
+    Return:
+        Processors list[processor]: The list of quantum circuits as a Perceval processors
+    """
+
+    '''Circuit implementation'''
+    from qiskit.visualization import circuit_drawer
+    from itertools import combinations
+    num_qubits = len(pauli_string)
+    if method == "WFT":
+        n_occ = n_orbs * 2 # number of spin orbitals
+    elif method == "DFT":
+        n_occ = n_orbs // 2 # number of spatial orbitals
+    initial_circuits = []
+
+    for i in range(n_occ): initial_circuits += [QuantumCircuit(num_qubits)]
+
+
+
+    '''Intial states'''
+    for state in range(n_occ):  # binarystring representation of the integer
+        for i in list_of_ones(state, num_qubits):
+            initial_circuits[state].x(i)
+
+
+    circuits = [
+        TwoLocal(num_qubits, 'ry', 'cx', 'linear', reps=layers,
+                 initial_state=initial_circuits[state]) for state in range(n_occ)]
+
+    circuits = [c.assign_parameters(lp) for c in circuits]
+
+
+
+
+    intial_states = []
+    for c in range(len(initial_circuits)):
+        sv = Statevector.from_instruction(initial_circuits[c])
+        index = list(sv.data).index(1 + 0j)  # position of the "1"
+        bitstring = format(index, f'0{sv.num_qubits}b')
+        bits_list = [int(b) for b in bitstring]
+        intial_states.append(pcvl.LogicalState(bits_list))
+
+    # print('Printing circuit')
+    # print(circuits[0].decompose())
+
+
[…]
