### 2025-10-15 4b1f532
updated ansatz.py and an example for usage
