+++ qlass/vqe/ansatz.py
@@ -1,7 +1,7 @@
 from perceval.converters import QiskitConverter
 from perceval.utils import NoiseModel
 from qiskit import transpile, QuantumCircuit
-from qiskit.circuit.library import TwoLocal
+from qiskit.circuit.library import n_local
 from qiskit.quantum_info import Statevector
 import perceval as pcvl
 import numpy as np
@@ -76,8 +76,25 @@ def custom_unitary_ansatz(lp: np.ndarray,
 
 def list_of_ones(computational_basis_state: int, n_qubits):
     """
-    Indices of ones in the binary expansion of an integer in big endian
-    order. e.g. 010110 -> [1, 3, 4] (which is the reverse of the qubit ordering...)
+    Return the indices of ones in the binary expansion of an integer, in big-endian order.
+
+    Parameters
+    ----------
+    computational_basis_state : int
+        Integer representing the computational basis state.
+    n_qubits : int
+        Total number of qubits (length of the binary string).
+
+    Returns
+    -------
+    list of int
+        List of indices where the binary representation has ones.
+        Indices are in big-endian order. For example, for `n_qubits=6` and
+        `computational_basis_state=22` (binary `010110`), the result is `[1, 3, 4]`.
+
+    Notes
+    -----
+    The output index ordering corresponds to the big-endian representation of the qubits.
     """
 
     bitstring = format(computational_basis_state, 'b').zfill(n_qubits)
@@ -86,48 +103,87 @@ def list_of_ones(computational_basis_state: int, n_qubits):
 
 
 
-def hf_ansatz(layers: int, n_orbs: int, lp: np.ndarray, pauli_string: str, method: str, noise_model: NoiseModel = None) -> Processor:
+def hf_ansatz(layers: int, n_orbs: int, lp: np.ndarray, pauli_string: str, method: str, cost = "VQE", noise_model: NoiseModel = None) -> Processor:
     """
-    Args:
-        layers (int): number of circuit layers
-        n_orbs (int): number of orbitals
-        lp (np.ndarray): array of parameter values
-        pauli_string (str): Pauli string
-        method (str): 'WFT' n_occ spin orbitals mapped into n_aubits, 'DFT' spatial orbitals mapped into log2(N) qubits by considering only spin-alpha or spin-beta blocks
-        see article https://scipost.org/SciPostPhys.14.3.055 for DFT mapping.
-        noise_model (NoiseModel): A perceval NoiseModel object representing the noise model
-    Return:
-        Processors list[processor]: The list of quantum circuits as a Perceval processors
-    """
+        Build a Hartree–Fock-based variational ansatz using Qiskit's ``n_local`` circuit,
+        combined with initial reference states and compiled into Perceval processors.
+
+        Parameters
+        ----------
+        layers : int
+            Number of circuit layers (repetitions) in the ansatz.
+        n_orbs : int
+            Number of orbitals.
+        lp : np.ndarray
+            Array of parameter values for the ansatz circuit.
+        pauli_string : str
+            Pauli operator string defining the measurement basis.
+        method : str
+            Mapping method. One of:
+                - ``"WFT"`` : Wavefunction theory mapping (spin orbitals → qubits)
+                - ``"DFT"`` : Density functional mapping (spatial orbitals → qubits)
+        cost : str, optional
+            Type of cost function to prepare. One of:
+                - ``"VQE"`` : Return only the ground-state processor (default)
+                - ``"e-VQE"`` : Return a list of processors for excited states
+        noise_model : NoiseModel, optional
[…]
