### 2025-10-15 3c72a43
Update ansatz.py
