### 2025-10-15 f8a82e1
Update paper.md
Update affiliation as per request in pre-review to JOSS
