### 2025-10-15 unitaryfoundation/qlass#121 @nathanshammah
Update references to paper.md
Update references to paper.md to include DOIs, as detailed in the thread with comments by the editorial bot.
https://github.com/openjournals/joss-reviews/issues/9079#issuecomment-3334517170
Once passing, ping the bot in that thread.
