### 2025-10-15 c255684
Update paper.bib
fix typo (DIO) in paper.bib
