+++ paper/paper.bib
@@ -71,7 +71,7 @@
   pages={931},
   year={2023},
   publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
-  DIO={10.22331/q-2023-02-21-931}
+  DOI={10.22331/q-2023-02-21-931}
 }
 
 @article{killoran2019strawberry,
