### 2025-10-15 #9079 @danielskatz
> [@danielskatz](https://github.com/danielskatz) apologies for the slow reply. I am a coauthor on the paper. We were checking the review status page on https://joss.theoj.org/papers/a74a8467e1dc986b0d4c1e1a64654f2a, where it read that the review process did not start yet and we thought there were no communications to respond to, and we did not see this thread from github notifications. Apologies for this, we will work on this with high priority.

A JOSS review has two parts: pre-review, where checks are made, an editor is assigned, and reviewers are assigned, as well the review, and authors can be called upon in both.

In any case, thanks for your changes, which I'll look at now
