### 2025-10-16 59125a3
Update Hamiltonian.py with Hchain KS hamiltonian
