+++ qlass/quantum_chemistry/hamiltonians.py
@@ -224,7 +224,7 @@ def LiH_hamiltonian(R=1.5, charge=0, spin=0, num_electrons=2, num_orbitals=2) ->
     Returns:
         Dict[str, float]: Hamiltonian dictionary with Pauli string keys
     """
-    
+
     # Create geometry in OpenFermion format
     geometry = [('Li', (0.0, 0.0, 0.0)), ('H', (0.0, 0.0, R))]
     
@@ -356,3 +356,152 @@ def LiH_hamiltonian_tapered(R: float) -> Dict[str, float]:
     qubit_op = symmetry_conserving_bravyi_kitaev(fermionic_op, active_spin_orbitals, active_n_elec)
 
     return sparsepauliop_dictionary(qubit_op)
+
+def Hchain_KS_hamiltonian(n_hydrogens=2, R=1.2):
+    """
+        Generate the one-body Hamiltonian for a linear chain of hydrogen atoms at a given bond length.
+
+        This function constructs a non-interacting one-body Hamiltonian for a chain of hydrogen atoms
+        using Density Functional Theory (DFT) with a closed-shell Hartree-Fock (RHF) method.
+        The hydrogen atoms are placed linearly along the z-axis with a uniform bond length ``R``.
+        The resulting molecular orbitals are transformed into a qubit Hamiltonian representation.
+
+        Parameters
+        ----------
+        n_hydrogens : int, optional
+            Number of hydrogen atoms in the linear chain. Must be an even integer.
+            Default is 2.
+        R : float, optional
+            Bond length between adjacent hydrogen atoms in angstroms. Default is 1.2.
+
+        Returns
+        -------
+        H_qubit_dic : dict
+            Dictionary representation of the qubit Hamiltonian in terms of Pauli operators.
+            The keys correspond to Pauli strings and the values are their coefficients.
+        mo_energy : list of float
+            Molecular orbital energies computed from PySCF.
+        n_occupied : int
+            Number of occupied orbitals in the closed-shell configuration.
+
+        Notes
+        -----
+        - The electronic structure is calculated using the minimal ``sto-3g`` basis set.
+        - The function internally performs the following steps:
+            1. Builds the molecular geometry.
+            2. Runs RHF self-consistent field (SCF) calculation via PySCF.
+            3. Constructs the Fock and overlap matrices in the atomic orbital (AO) basis.
+            4. Transforms to an orthogonalized atomic orbital (OAO) basis.
+            5. Maps the resulting Hamiltonian to a qubit representation.
+        - The transformation to the qubit Hamiltonian uses a helper function
+          ``transformation_Hmatrix_Hqubit`` and a dictionary builder ``sparsepauliop_dictionary``.
+
+        """
+    from pyscf import gto, scf
+    geometry = []
+    numberof_qubits = int(np.log2(n_hydrogens))
+
+    for d in range(n_hydrogens // 2):
+        geometry.append(('H', (0.0, 0.0, - (R / 2. + d * R))))
+        geometry.append(('H', (0.0, 0.0, + (R / 2. + d * R))))
+
+    molecule = gto.M(atom=geometry, basis='sto-3g')
+    mf = scf.RHF(molecule)
+    mf.scf()
+    dm = mf.make_rdm1()
+    h1e = mf.get_hcore()
+    veff = mf.get_veff()
+    F_AO = mf.get_fock()
+    S_AO = mf.get_ovlp()
+    # Compute the inverse square root of the overlap matrix S
+    S_eigval, S_eigvec = np.linalg.eigh(S_AO)
+    S_sqrt_inv = S_eigvec @ np.diag((S_eigval) ** (-1. / 2.)) @ S_eigvec.T
[…]
