+++ qlass/quantum_chemistry/__init__.py
@@ -11,6 +11,8 @@ from .hamiltonians import (
     LiH_hamiltonian, 
     generate_random_hamiltonian, 
     LiH_hamiltonian_tapered,
+    Hchain_KS_hamiltonian,
+    transformation_Hmatrix_Hqubit,
     pauli_commute,
     group_commuting_pauli_terms,
     group_commuting_pauli_terms_openfermion_hybrid,
@@ -26,6 +28,8 @@ __all__ = [
     "LiH_hamiltonian",
     "generate_random_hamiltonian", 
     "LiH_hamiltonian_tapered",
+    "Hchain_KS_hamiltonian"
+    "transformation_Hmatrix_Hqubit",
     "pauli_commute",
     "group_commuting_pauli_terms",
     "group_commuting_pauli_terms_openfermion_hybrid",
