### 2025-10-16 a5dc181
Update Hchain KS hamiltonian usage with an example
