### 2025-10-16 bbf0852
add necessary utility functions for computing energy of post-selected photonic unitary
