+++ qlass/utils/utils.py
@@ -423,3 +423,222 @@ def loss_function_matrix(
         loss += coefficient * expectation
     
     return float(np.real(loss))
+
+def permanent(matrix: np.ndarray) -> complex:
+    """
+    Calculate the permanent of a matrix.
+    
+    The permanent is like a determinant but without alternating signs:
+    Perm(A) = Σ_{σ∈Sₘ} Π_{i=1 to m} A_{i,σ(i)}
+    
+    Parameters:
+    -----------
+    matrix : np.ndarray
+        Square matrix to calculate permanent of
+        
+    Returns:
+    --------
+    complex
+        The permanent of the matrix
+    """
+    from itertools import permutations
+    
+    n = matrix.shape[0]
+    if n == 0:
+        return 1
+    
+    perm = 0
+    for sigma in permutations(range(n)):
+        product = 1
+        for i in range(n):
+            product *= matrix[i, sigma[i]]
+        perm += product
+    
+    return perm
+
+
+def logical_state_to_modes(logical_state: int, m: int) -> List[int]:
+    """
+    Convert a logical qubit state to the set of occupied photon modes.
+    
+    Parameters:
+    -----------
+    logical_state : int
+        Integer representing the logical state (0 to 2^m - 1)
+    m : int
+        Number of qubits
+        
+    Returns:
+    --------
+    List[int]
+        List of occupied mode indices (0-indexed)
+    """
+    # Convert integer to bit list
+    bits = [(logical_state >> (m - 1 - k)) & 1 for k in range(m)]
+    
+    # For qubit k (0-indexed), the modes are 2k and 2k+1
+    # |0⟩ₖ → mode 2k, |1⟩ₖ → mode 2k+1
+    modes = []
+    for k in range(m):
+        mode = 2 * k + bits[k]
+        modes.append(mode)
+    
+    return modes
+
+
+def photon_to_qubit_unitary(U_photon: np.ndarray) -> np.ndarray:
+    """
+    Convert a photon unitary to the effective qubit unitary via post-selection.
+    
+    Parameters:
+    -----------
+    U_photon : np.ndarray
+        The 2m × 2m unitary matrix acting on photon modes
+        
+    Returns:
+    --------
+    np.ndarray
+        The 2^m × 2^m effective qubit unitary matrix
+    """
+    # Determine number of qubits
+    modes_2m = U_photon.shape[0]
+    if modes_2m % 2 != 0:
+        raise ValueError("Photon unitary must have even dimension (2m × 2m)")
+    
+    m = modes_2m // 2
+    num_logical_states = 2 ** m
+    
+    # Initialize the qubit unitary
+    U_qubit = np.zeros((num_logical_states, num_logical_states), dtype=complex)
+    
+    # For each pair of input and output logical states
+    for r in range(num_logical_states):
+        for s in range(num_logical_states):
+            # Get the mode sets for input state |r⟩ and output state |s⟩
+            R = logical_state_to_modes(r, m)
+            S = logical_state_to_modes(s, m)
+            
+            # Extract the submatrix U(S,R)
+            # Rows indexed by S, columns indexed by R
+            submatrix = U_photon[np.ix_(S, R)]
+            
+            # The matrix element is the permanent of this submatrix
+            U_qubit[s, r] = permanent(submatrix)
+    
+    return U_qubit
+
+
+def compute_energy_postselected(
+    U_photon: np.ndarray,
+    initial_state_logical: np.ndarray,
+    hamiltonian_terms: List[Tuple[float, str]]
+) -> Tuple[float, float, np.ndarray]:
+    """
+    Compute the energy of a post-selected state after photonic evolution.
+    
+    The process is:
[…]
