### 2025-10-16 88bb5cc
Update qlass/quantum_chemistry/hamiltonians.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
