### 2025-10-17 92573c1
update VQE to use photonic_unitary
