### 2025-10-17 e7466f9
update VQE class to work with photonic_unitary and added relevant tests and example file
