### 2025-10-17 bf85676
Update qlass/vqe/vqe.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
