### 2025-10-17 ae620ca
Update qlass/utils/utils.py
Use Ryser's formula for permanent instead of brute force version.

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
