### 2025-10-17 af861ef
fix bug in permanent() function
