### 2025-10-17 #9079 @nathanshammah
Thank you @danielskatz. Here is a list of suggested referees:

- Maxime Garnier, INRIA (Graphix dev), gh: mgarnier59
- Zoltan Zimboras, University of Helsinki (Piquasso dev), gh: zimboras
- Jean Senellart, Quandela (Perceval dev), France, gh: jsenellart
- Josh Izaac, Xanadu Inc (PennyLane dev),  gh: josh146
- Mauro D'Arcangelo, PASQAL, gh: darcangelomauro

I can provide emails privately.
