### 2025-10-17 18f62fd
Ensemble VQE for excited states
