+++ qlass/tests/test_vqe.py
@@ -12,6 +12,7 @@ from qlass.vqe import (
 
 from qlass.quantum_chemistry import (
     LiH_hamiltonian,
+    Hchain_KS_hamiltonian
 )
 
 import warnings
@@ -34,6 +35,7 @@ def simple_vqe():
     num_params = 4
     return VQE(hamiltonian=hamiltonian, executor=mock_executor, num_params=num_params)
 
+
 def test_vqe_init(simple_vqe):
     """Tests if the VQE class is initialized with the correct attributes."""
     assert simple_vqe.num_qubits == 2
@@ -46,7 +48,7 @@ def test_vqe_run(simple_vqe):
     """Tests the main `run` method to ensure optimization completes."""
     # Run with a small number of iterations for speed
     final_energy = simple_vqe.run(max_iterations=10, verbose=False)
-    
+
     assert isinstance(final_energy, float)
     assert simple_vqe.optimization_result is not None
     # The callback should populate the history
@@ -54,6 +56,7 @@ def test_vqe_run(simple_vqe):
     assert len(simple_vqe.parameter_history) > 0
     # The final energy should match the one in the result object
     assert np.isclose(final_energy, simple_vqe.optimization_result.fun)
+
 
 def test_get_optimal_parameters(simple_vqe):
     """
@@ -65,9 +68,9 @@ def test_get_optimal_parameters(simple_vqe):
         simple_vqe.get_optimal_parameters()
 
     # After running
-    simple_vqe.run(max_iterations=10, verbose=False)
+    simple_vqe.run(max_iterations=3, verbose=False)
     optimal_params = simple_vqe.get_optimal_parameters()
-    
+
     assert isinstance(optimal_params, np.ndarray)
     assert len(optimal_params) == simple_vqe.num_params
     assert np.allclose(optimal_params, simple_vqe.optimization_result.x)
@@ -79,16 +82,16 @@ def test_compare_with_exact(simple_vqe):
     # Should raise error before running
     with pytest.raises(ValueError, match="VQE optimization has not been run yet."):
         simple_vqe.compare_with_exact(0.0)
-    
+
     simple_vqe.run(max_iterations=10, verbose=False)
     vqe_energy = simple_vqe.optimization_result.fun
     exact_energy = 0.5
-    
+
     comparison = simple_vqe.compare_with_exact(exact_energy)
-    
+
     expected_abs_error = abs(vqe_energy - exact_energy)
     expected_rel_error = expected_abs_error / abs(exact_energy)
-    
+
     assert comparison['vqe_energy'] == vqe_energy
     assert comparison['exact_energy'] == exact_energy
     assert np.isclose(comparison['absolute_error'], expected_abs_error)
@@ -102,10 +105,10 @@ def test_vqe_pipeline():
         sampler = Sampler(processor)
         samples = sampler.samples(10_000)
         return samples
-    
+
     # Number of qubits
     num_qubits = 2
-    
+
     # Generate a 2-qubit Hamiltonian
     hamiltonian = LiH_hamiltonian(num_electrons=2, num_orbitals=1)
 
@@ -115,11 +118,44 @@ def test_vqe_pipeline():
         executor=executor,
         num_params=2*num_qubits, # Number of parameters in the linear entangled ansatz
     )
-    
+
     # Run the VQE optimization
     vqe_energy = vqe.run(
         max_iterations=10,
         verbose=True
+    )
+
+    if not isinstance(vqe_energy, float):
+        raise ValueError("Optimization result is not a valid float")
+
+def test_evqe_pipeline():
+
+    # Define an executor function that uses the linear entangled ansatz
[…]
