+++ qlass/utils/utils.py
@@ -423,3 +423,225 @@ def loss_function_matrix(
         loss += coefficient * expectation
     
     return float(np.real(loss))
+
+def e_vqe_loss_function(lp: np.ndarray, H: Dict[str, float], executor, energy_collector, weight_option: str= "weighted") -> float:
+    """
+    Compute the loss function for the ensemble Variational Quantum Eigensolver (VQE)
+    with automatic Pauli grouping for measurement optimization.
+
+    This function automatically groups commuting Pauli terms to reduce the number
+    of measurements required. The grouping is applied transparently without
+    changing the function interface, providing automatic optimization for ensemble VQE
+    algorithms.
+
+    The function works with executors that return either:
+    1. Fock states (from linear optical circuits) - exqalibur.FockState objects
+    2. Bitstring tuples (from regular qubit-based circuits)
+    3. Bitstring strings (from Qiskit Sampler)
+
+    The executor should return samples in one of these formats:
+    - Dict with 'results' key: {'results': [samples]}
+    - Direct list of samples: [samples]
+    - Qiskit-style format with bitstrings or counts
+
+    Parameters
+    ----------
+    lp : np.ndarray
+        Array of variational circuit parameters. Typically optimized to minimize
+        the expectation value of the Hamiltonian.
+    H : dict of {str: float}
+        Hamiltonian represented as a dictionary mapping Pauli strings
+        (e.g., ``'X0 Z1'``) to their coefficients.
+    executor : callable
+        Function or callable object that executes the quantum circuit and returns
+        measurement samples. Must accept arguments ``(lp, pauli_string)`` and return
+        samples in one of the accepted formats.
+    energy_collector : object
+        Object responsible for tracking or logging the energy convergence history.
+        Must implement a method ``energies_convergence(energies, n_ensembles, total_loss)``.
+    weight_option : {'weighted', 'uniform'}
+        Scheme for assigning ensemble weights:
+        - ``'weighted'`` (default): Linearly decreasing weights with index, i.e., w_i < w_j for i > j.
+        - ``'equi'`` : Equal weights for all occupied orbitals, w_i = w_j.
+        - ``'ground_state_only'`` : Only the ground state contributes, w_0 = 1, others 0.
+
+    Returns
+    -------
+    loss : float
+        The computed ensemble VQE loss value, equal to the weighted sum of ensemble
+        energies.
+    """
+    # Import here to avoid circular imports
+    try:
+        from qlass.quantum_chemistry.hamiltonians import group_commuting_pauli_terms
+        use_grouping = True
+    except ImportError:
+        # Fallback to individual processing if grouping not available
+        use_grouping = False
+
+    loss = 0.0
+    lst_energies = None
+
+    if use_grouping:
+        # Use automatic grouping for optimized measurements
+        grouped_hamiltonians = group_commuting_pauli_terms(H)
+
+        for group in grouped_hamiltonians:
+            # Each group contains mutually commuting terms
+            # In the future, this could be optimized to measure entire groups simultaneously
+            # For now, we process each term individually but with the grouping organization
+            for pauli_string, coefficient in group.items():
+
+                samples = executor(lp, pauli_string)
+
+                # Handle different executor return formats
+                sample_lists = [_extract_samples_from_executor_result(s) for s in samples]
+
+                # Normalize samples to consistent format
+                normalized_samples = [normalize_samples(sample_list) for sample_list in sample_lists]
+
+                prob_dist = [get_probabilities(normalized_sample) for normalized_sample in normalized_samples]
+                pauli_bin = pauli_string_bin(pauli_string)
+
+                qubit_state_marg = [qubit_state_marginal(pd) for pd in prob_dist]
[…]
