### 2025-10-17 a4b1696
Ensemble VQE for excited states
