### 2025-10-21 d6aa668
Ensemble VQE for excited states resolving conflit
