+++ qlass/tests/test_vqe.py
@@ -1,3 +1,4 @@
+
 import numpy as np
 import perceval as pcvl
 from perceval.algorithm import Sampler
@@ -35,7 +36,6 @@ def simple_vqe():
     num_params = 4
     return VQE(hamiltonian=hamiltonian, executor=mock_executor, num_params=num_params)
 
-
 def test_vqe_init(simple_vqe):
     """Tests if the VQE class is initialized with the correct attributes."""
     assert simple_vqe.num_qubits == 2
@@ -57,7 +57,6 @@ def test_vqe_run(simple_vqe):
     # The final energy should match the one in the result object
     assert np.isclose(final_energy, simple_vqe.optimization_result.fun)
 
-
 def test_get_optimal_parameters(simple_vqe):
     """
     Tests that `get_optimal_parameters` returns correct parameters after a run
@@ -68,7 +67,7 @@ def test_get_optimal_parameters(simple_vqe):
         simple_vqe.get_optimal_parameters()
 
     # After running
-    simple_vqe.run(max_iterations=3, verbose=False)
+    simple_vqe.run(max_iterations=10, verbose=False)
     optimal_params = simple_vqe.get_optimal_parameters()
 
     assert isinstance(optimal_params, np.ndarray)
@@ -160,6 +159,7 @@ def test_evqe_pipeline():
 
     if not isinstance(vqe_energy, float):
         raise ValueError("Optimization result is not a valid float")
+
 
 def test_custom_unitary_ansatz():
     """
@@ -295,10 +295,10 @@ def test_vqe_init_with_unitary_executor():
         hamiltonian=hamiltonian,
         executor=identity_unitary_executor,
         num_params=2,
-        executor_type="unitary"
+        executor_type="qubit_unitary"
     )
 
-    assert vqe.executor_type == "unitary"
+    assert vqe.executor_type == "qubit_unitary"
     assert vqe.num_qubits == 2
     assert vqe.num_params == 2
 
@@ -350,7 +350,7 @@ def test_vqe_compare_unitary_vs_sampling():
         hamiltonian=hamiltonian,
         executor=unitary_exec,
         num_params=2,
-        executor_type="unitary"
+        executor_type="qubit_unitary"
     )
 
     vqe_sampling = VQE(
@@ -375,3 +375,169 @@ def test_vqe_compare_unitary_vs_sampling():
     # Expected: 0.5 * 1 + 1.0 * 1 = 1.5
     assert np.isclose(energy_unitary, 1.5)
     assert np.isclose(energy_sampling, 1.5, atol=0.01)
+
+def photonic_identity_executor(params):
+    """Photonic unitary executor that returns identity."""
+    return np.eye(4, dtype=complex)  # 2 qubits, 4 modes
+
+
+def photonic_parametrized_executor(params):
+    """Photonic unitary executor with actual parameters."""
+    # Create a simple 4×4 unitary using rotation-like structure
+    theta = params[0]
+    phi = params[1] if len(params) > 1 else 0
+
+    # Simple parameterized unitary for 2 qubits (4 modes)
+    U = np.eye(4, dtype=complex)
+
+    # Apply rotation-like transformation
+    c, s = np.cos(theta), np.sin(theta)
+    U[0, 0] = c * np.exp(1j * phi)
+    U[0, 1] = -s
+    U[1, 0] = s
+    U[1, 1] = c * np.exp(-1j * phi)
+
+    return U
+
+
+def test_vqe_init_with_photonic_unitary_executor():
+    """Test VQE initialization with photonic_unitary executor type."""
+    hamiltonian = {"II": -0.5, "ZZ": 1.0}
+
+    vqe = VQE(
+        hamiltonian=hamiltonian,
+        executor=photonic_identity_executor,
[…]
