+++ qlass/vqe/vqe.py
@@ -6,27 +6,30 @@ import matplotlib.pyplot as plt
 from qlass.utils import loss_function, e_vqe_loss_function
 from qlass.utils.utils import DataCollector
 
+
 class VQE:
     """
     Variational Quantum Eigensolver for photonic quantum computing.
-    
+
     This class provides a high-level interface for running VQE experiments
     on photonic simulators, finding ground state energies of quantum systems.
     """
-    
+
     def __init__(
-        self, 
-        hamiltonian: Dict[str, float],
-        executor: Callable,
-        num_params: int,
-        optimizer: str = "COBYLA",
-        executor_type: str = 'sampling',
+            self,
+            hamiltonian: Dict[str, float],
+            executor: Callable,
+            num_params: int,
+            optimizer: str = "COBYLA",
+            executor_type: str = 'sampling',
+            initial_state: np.ndarray = None, # For now only relevant for photonic_unitary,
+
     ):
         """
         Initialize the VQE solver.
-        
+
         Args:
-            hamiltonian (Dict[str, float]): Hamiltonian dictionary with Pauli string keys 
+            hamiltonian (Dict[str, float]): Hamiltonian dictionary with Pauli string keys
                                            and coefficient values
             executor (Callable): Custom executor function, if None, a default one will be created
             num_params (int): number of parameters that the executor accepts
@@ -37,16 +40,16 @@ class VQE:
         self.executor = executor
         self.num_params = num_params
         self.optimizer = optimizer
-        
+        self.initial_state = initial_state
+
         # Extract number of qubits from the Hamiltonian
         self.num_qubits = len(next(iter(hamiltonian.keys())))
-        
+
         # Executor type for loss function computation
-        if executor_type in ["sampling", "unitary"]:
+        if executor_type in ["sampling", "qubit_unitary", "photonic_unitary"]:
             self.executor_type = executor_type
         else:
-            raise ValueError(f"Invalid executor_type: {executor_type}. Must be either sampling or unitary.")
-        
+            raise ValueError(f"Invalid executor_type: {executor_type}. Must be either sampling, qubit_unitary or photonic_unitary.")
 
         # Results storage
         self.optimization_result = None
@@ -54,16 +57,21 @@ class VQE:
         self.parameter_history = []
         self.loss_history = []
         self.energy_collector = DataCollector()
-    
+
     def _callback_vqe(self, params):
         """Callback function to record optimization progress."""
-        if self.executor_type == "unitary":
+        """Callback function to record optimization progress."""
+        if self.executor_type == "qubit_unitary":
             from qlass.utils import loss_function_matrix
             energy = loss_function_matrix(params, self.hamiltonian, self.executor)
+        elif self.executor_type == "photonic_unitary":
+            from qlass.utils import loss_function_photonic_unitary
+            energy = loss_function_photonic_unitary(
+                params, self.hamiltonian, self.executor, self.initial_state)
         else:
             from qlass.utils import loss_function
             energy = loss_function(params, self.hamiltonian, self.executor)
-        
+
         self.energy_history.append(energy)
         self.parameter_history.append(params.copy())
 
@@ -72,7 +80,7 @@ class VQE:
         cost = e_vqe_loss_function(params, self.hamiltonian, self.executor, self.energy_collector)
         self.loss_history.append(cost)
         self.parameter_history.append(params.copy())
-        
+
     def run(self, initial_params=None, max_iterations=100, verbose=True, weight_option: str = "weighted",
             cost: str = "VQE"):
         """
@@ -124,19 +132,19 @@ class VQE:
         # Initialize parameters if not provided
         if initial_params is None:
             initial_params = np.random.rand(self.num_params)
-        
+
         # Reset history
         self.energy_history = []
         self.parameter_history = []
-        
+
         if verbose:
[…]

[…]
