### 2025-10-21 3311543
Update qlass/vqe/vqe.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
