### 2025-10-21 09151cb
Update qlass/vqe/vqe.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
