### 2025-10-21 3d302a0
Update qlass/utils/utils.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
