### 2025-10-21 bcee557
Update examples/ensemble_vqe_H4.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
