### 2025-10-22 2be6462
update loss_function_photonic_unitary in to be matrix-free in utils.py
