### 2025-10-22 dbbd7de
revert initial_state back to being vector instead of integer
