### 2025-10-22 aaf21ce
Update qlass/utils/utils.py
improve succes_prob computation

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
