### 2025-10-22 5968c1f
Update qlass/utils/utils.py
improve term_expectation calculation

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
