### 2025-10-23 e888720
Ensemble VQE for excited states resolving conflit
