### 2025-10-23 e8c4132
Ensemble VQE for excited states resolving conflit
