### 2025-10-23 da7fcf9
Ensemble VQE for excited states resolving conflit
