+++ .github/workflows/docs-build.yml
@@ -28,7 +28,7 @@ jobs:
           make html
           
       - name: Upload documentation
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: documentation
           path: docs/_build/html/  
