### 2025-10-29 b49394c
update loss_function_photonic_unitary to accept ancilla modes and update VQE class accordingly (to work with postselected CNOTs)
