### 2025-10-30 6828cc2
update test_utils.py to cover new lines of code
