### 2025-10-30 162c0b3
adding an example script vqe_example_pu_ps.py for photonic unitary with post selection implemented
