### 2025-11-03 efa7788
update examples.rst
