+++ docs/examples.rst
@@ -61,50 +61,200 @@ The report provides insights into component counts, estimated photon loss, and o
 Variational Quantum Eigensolver (VQE)
 -------------------------------------
 
-Implementing VQE to find the ground state energy of a molecule:
+VQE with Sampling Executor
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Standard VQE using sampling-based executor with Hartree-Fock ansatz:
 
 .. code-block:: python
 
-   import warnings
-   warnings.simplefilter('ignore')
-   warnings.filterwarnings('ignore')
-
    import numpy as np
-   from perceval.algorithm import Sampler
-   
-   from qlass.vqe import VQE, le_ansatz
+   from qlass.vqe import VQE
    from qlass.quantum_chemistry import LiH_hamiltonian, brute_force_minimize
+   from qlass.vqe.ansatz import hf_ansatz
+   from perceval.algorithm import Sampler
+   import warnings
+   warnings.filterwarnings('ignore')
    
-   # Create a molecular Hamiltonian
-   hamiltonian = LiH_hamiltonian(num_electrons=2, num_orbitals=1)
+   # Create molecular Hamiltonian
+   ham = LiH_hamiltonian(num_electrons=2, num_orbitals=1)
    
-   # Define an executor function that runs the quantum circuit
+   # Define executor using Hartree-Fock ansatz
    def executor(params, pauli_string):
-       processor = le_ansatz(params, pauli_string)
-       sampler = Sampler(processor)
-       samples = sampler.samples(10000)
+       processors = hf_ansatz(1, 1, params, pauli_string, method="WFT", cost="VQE")
+       samplers = Sampler(processors)
+       samples = samplers.samples(10_000)
        return samples
    
-    # Initialize the VQE solver with the custom executor
-    vqe = VQE(
-        hamiltonian=hamiltonian,
-        executor=executor,
-        num_params=4, # Number of parameters in the linear entangled ansatz
-    )
-    
-    # Run the VQE optimization
-    vqe_energy = vqe.run(
-        max_iterations=10,
-        verbose=True
-    )
+   # Initialize VQE solver
+   vqe = VQE(
+       hamiltonian=ham,
+       executor=executor,
+       num_params=4,
+   )
    
-   # Print the results
+   # Run optimization
+   vqe_energy = vqe.run(max_iterations=50, verbose=True)
+   
+   # Compare with exact solution
+   exact_energy = brute_force_minimize(ham)
    print(f"VQE Energy: {vqe_energy:.6f}")
-   
-   # Compare with the exact solution
-   exact_energy = brute_force_minimize(hamiltonian)
    print(f"Exact Energy: {exact_energy:.6f}")
-   print(f"Energy Difference: {abs(vqe_energy - exact_energy):.6f}")
+
+VQE with Qubit Unitary Executor
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Using unitary matrices directly for VQE:
+
+.. code-block:: python
+
+   import numpy as np
+   from qlass.vqe import VQE
+   from qlass.quantum_chemistry import LiH_hamiltonian, brute_force_minimize
+   from scipy.linalg import expm
+   
+   # Define unitary executor using parameterized generators
+   def unitary_executor(params):
+       num_qubits = 2
+       dim = 2**num_qubits
+       
+       # Define Pauli matrices
+       I = np.eye(2)
+       X = np.array([[0, 1], [1, 0]])
+       Y = np.array([[0, -1j], [1j, 0]])
[…]
