### 2025-11-03 f53a08a
update examples.rst
