### 2025-11-03 c963d33
update qlass.vqe.rst
