### 2025-11-03 4218c36
update qlass.vqe.rst
