### 2025-11-03 b6eb7cb
update qlass.compiler.rst
