### 2025-11-05 90c4394
update hardware_config.py and compiler.py
