### 2025-11-05 10f9071
Update qlass/compiler/hardware_config.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
