### 2025-11-05 f0bacfc
Update qlass/vqe/vqe.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
