+++ qlass/vqe/vqe.py
@@ -34,7 +34,10 @@ class VQE:
             executor (Callable): Custom executor function, if None, a default one will be created
             num_params (int): number of parameters that the executor accepts
             optimizer (str): Optimization method to use. Any method supported by scipy.optimize.minimize
-            executor_type (str): Type of executor - "sampling", "qubit_unitary", "photonic_unitary"
+            executor_type (str): Type of executor. 
+                - "sampling": Uses measurement-based sampling.
+                - "qubit_unitary": For qubit-based simulations (e.g., standard circuit-based quantum computing).
+                - "photonic_unitary": For photonic quantum computing, where the initial state and ancillary modes may be relevant for post-selection. Use this when simulating photonic systems or when post-selection on ancillary modes is required.
             ancillary_modes (List[int], optional): List of ancillary mode indices
                 for post-selection when using 'photonic_unitary' executor.
         """
