### 2025-11-10 2b3b18d
Add plot_energy_landscape method to VQE class for energy visualization
