### 2025-11-10 2c0af00
Add test for plot_energy_landscape method in VQE class
