### 2025-11-10 c04438e
Generalize plot_energy_landscape method to support all executor types
