### 2025-11-11 bf68eaf
updated all the loss functions to also return contributions of each hamiltonian term
