+++ qlass/tests/test_utils.py
@@ -106,20 +106,20 @@ def test_loss_function_automatic_grouping():
     # test case 1: simple 2-qubit hamiltonian with commuting terms
     simple_ham = {"II": 1.0, "ZI": 0.5, "IZ": -0.3, "ZZ": 0.2}
     test_params = np.array([0.1, 0.2])
+
+    energy, contributions = loss_function(test_params, simple_ham, mock_executor)
+    assert isinstance(energy, float), "loss_function should return a float energy"
+    assert isinstance(contributions, dict), "loss_function should return a contrib dict"
     
-    # Function should work with automatic grouping
-    result = loss_function(test_params, simple_ham, mock_executor)
-    assert isinstance(result, float), "loss_function should return a float"
-    
-    # test case 2: empty hamiltonian should work
     empty_ham = {}
-    result_empty = loss_function(test_params, empty_ham, mock_executor)
-    assert result_empty == 0.0, "Empty Hamiltonian should give zero loss"
-    
-    # test case 3: single term should work
+    energy_empty, contrib_empty = loss_function(test_params, empty_ham, mock_executor)
+    assert energy_empty == 0.0, "Empty Hamiltonian should give zero loss"
+    assert contrib_empty == {}, "Empty Hamiltonian should give empty contribs"
+
     single_ham = {"ZZ": 1.0}
-    result_single = loss_function(test_params, single_ham, mock_executor)
-    assert isinstance(result_single, float), "Single term should work correctly"
+    energy_single, contrib_single = loss_function(test_params, single_ham, mock_executor)
+    assert isinstance(energy_single, float), "Single term should work correctly"
+    assert "ZZ" in contrib_single
 
 
 def test_loss_function_perceval_format():
@@ -128,15 +128,21 @@ def test_loss_function_perceval_format():
     def mock_perceval_executor(params, pauli_string):
         return {
             'results': [
-                pcvl.BasicState([1, 0, 1, 0]),  # |01⟩ 
+                pcvl.BasicState([1, 0, 1, 0]),  # |00⟩ 
                 pcvl.BasicState([0, 1, 0, 1]),  # |11⟩
-                pcvl.BasicState([1, 0, 0, 1]),  # |00⟩
+                pcvl.BasicState([1, 0, 0, 1]),  # |01⟩
             ]
         }
     
     hamiltonian = {"II": 0.5, "ZZ": 0.3}
-    result = loss_function(np.array([0.1, 0.2]), hamiltonian, mock_perceval_executor)
-    assert isinstance(result, float)
+    energy, contributions = loss_function(np.array([0.1, 0.2]), hamiltonian, mock_perceval_executor)
+    assert isinstance(energy, float)
+    assert isinstance(contributions, dict)
+    # Manual check: States are |00⟩, |11⟩, |01⟩. All 1/3 prob.
+    # II: 0.5 * 1.0 = 0.5
+    # ZZ: 0.3 * ( (1/3)*(+1) + (1/3)*(+1) + (1/3)*(-1) ) = 0.3 * (1/3) = 0.1
+    # Expected: 0.5 + 0.1 = 0.6
+    assert np.isclose(energy, 0.6)
 
 def test_loss_function_qiskit_bitstring_format():
     
@@ -145,8 +151,13 @@ def test_loss_function_qiskit_bitstring_format():
         return {'results': ['00', '01', '10', '11']}
     
     hamiltonian = {"II": 1.0, "ZZ": 0.5}
[…]
