### 2025-11-11 f91c754
add test function for plot_hamiltonian_contributions function
