### 2025-11-13 9a394b9
🔧 do not upper cap Python
see https://iscinumpy.dev/post/bound-version-constraints/ and related links
