### 2025-11-13 91f06c0
🔧 bump minimum supported Python version to 3.10
3.9 has reached end of life
