### 2025-11-13 f68fcd1
📄 proper SPDX license specifier
