### 2025-11-13 d9442ec
🔧 switch to using dependency groups
