### 2025-11-13 625fe65
🏷️ mark package as typed
