### 2025-11-13 3c0a553
🩹 fix project version in various places
