+++ src/qlass/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
 ]
 
 # Version information
-__version__ = "0.1.1"
+__version__ = "0.1.4"
