### 2025-11-13 6c97dfb
🔧 add proper pytest config
