+++ pyproject.toml
@@ -36,8 +36,11 @@ license-files = ["LICENSE"]
 [dependency-groups]
 test = [
     "pytest>=8.3.5",
+    "pytest-cov>=6.1.1",
     "pytest-mock>=3.15.1",
-    "coverage>=7.11.3"
+]
+dev = [
+    {include-group = "test"},
 ]
 
 [build-system]
@@ -47,3 +50,14 @@ build-backend = "uv_build"
 [project.urls]
 Homepage = "https://github.com/unitaryfoundation/qlass"
 Issues = "https://github.com/unitaryfoundation/qlass/issues"
+
+[tool.pytest.ini_options]
+minversion = "7.2"
+addopts = [
+    "-ra",
+    "--strict-markers",
+    "--strict-config",
+]
+xfail_strict = true
+log_cli_level = "INFO"
+testpaths = ["tests"]
