### 2025-11-13 57dd070
🔧 coverage configuration in pyproject.toml
