### 2025-11-13 f85b984
🔧 clean up dependency configuration
