### 2025-11-13 379ea19
✏️ fix typos in paper
