### 2025-11-13 a1649ba
👷 modernize documentation setup
