+++ .github/workflows/docs-build.yml
@@ -9,23 +9,13 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v5
-      
-      - name: Set up Python
-        uses: actions/setup-python@v6
-        with:
-          python-version: '3.12'  # Specify your Python version
-          
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install sphinx sphinx sphinx_rtd_theme # Add any Sphinx extensions you need
-          pip install -e .  # Install your package in development mode
+
+      - name: Install the latest version of uv
+        uses: astral-sh/setup-uv@v7
           
       - name: Build documentation
         run: |
-          sphinx-apidoc -o docs qlass
-          cd docs  
-          make html
+          uv run -p 3.12 --no-dev --group docs -m sphinx -T -b html -d docs/_build/doctrees docs docs/_build/html
           
       - name: Upload documentation
         uses: actions/upload-artifact@v5
