+++ pyproject.toml
@@ -41,6 +41,10 @@ test = [
     "pytest-cov>=6.1.1",
     "pytest-mock>=3.15.1",
 ]
+docs = [
+    "sphinx>=7",
+    "sphinx-rtd-theme>=3",
+]
 dev = [
     {include-group = "test"},
 ]
