### 2025-11-13 31fccec
🎨 use importlib metadata to set package version
