+++ pyproject.toml
@@ -2,7 +2,7 @@
 
 [project]
 name = "qlass"
-version = "0.1.4"
+version = "0.2.0"
 description = "VQE on linear optical circuits"
 authors = [{name = "Farrokh Labib", email = "farrokh@unitary.foundation"}]
 requires-python = ">=3.9,<3.14"
