### 2025-11-14 af1c5b7
Update conf.py
update version
