### 2025-11-15 unitaryfoundation/qlass#141 @goerz
Undefined `Sampler` in README example
The code snippet in the ["Variational Quantum Eigensolver (VQE)"](https://github.com/unitaryfoundation/qlass?tab=readme-ov-file#variational-quantum-eigensolver-vqe) section of the README references a `Sampler` that does not exist.

I would recommend verifying the code snippets in your README as part of the CI tests.

Ref. https://github.com/openjournals/joss-reviews/issues/9316
