### 2025-11-17 88d1155
Bump qiskit from 2.1.2 to 2.2.3
Bumps [qiskit](https://github.com/Qiskit/qiskit) from 2.1.2 to 2.2.3.
- [Release notes](https://github.com/Qiskit/qiskit/releases)
- [Changelog](https://github.com/Qiskit/qiskit/blob/main/docs/release_notes.rst)
- [Commits](https://github.com/Qiskit/qiskit/compare/2.1.2...2.2.3)

---
updated-dependencies:
- dependency-name: qiskit
  dependency-version: 2.2.3
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
