+++ pyproject.toml
@@ -8,7 +8,7 @@ authors = [{name = "Farrokh Labib", email = "farrokh@unitary.foundation"}]
 requires-python = ">=3.9,<3.14"
 
 dependencies = [
-"qiskit==2.1.2",
+"qiskit==2.2.3",
 "qiskit_aer",
 "numpy",
 "perceval-quandela==0.13.2",
