### 2025-11-17 unitaryfoundation/qlass#145 @FarLab
Update documentation
The documentation seems to be slightly outdated and probably needs to be updated.

(from joss review: https://github.com/openjournals/joss-reviews/issues/9316#issuecomment-3530069168)
