### 2025-11-17 unitaryfund/qlass#146 @FarLab
Add code formatter or linter
The project is not using a code formatter or linter at the moment. Consider looking into ruff, which is just excellent and very easy to configure. https://docs.astral.sh/ruff/ See https://github.com/munich-quantum-toolkit/core/blob/8c9f6ab24968401e450812fc0ff7d05b5ae07a63/.pre-commit-config.yaml#L67-L73 for some example configuration on how to run it and https://github.com/munich-quantum-toolkit/core/blob/8c9f6ab24968401e450812fc0ff7d05b5ae07a63/pyproject.toml#L204-L249 for some inspiration for configuration.

(from Joss review: https://github.com/openjournals/joss-reviews/issues/9316#issuecomment-3530069168)
