### 2025-11-17 unitaryfund/qlass#147 @FarLab
Add type checking
The project currently does not use type checking. Consider adding mypy and running it as part of CI (e.g. as part of pre-commit.ci). See https://github.com/munich-quantum-toolkit/core/blob/8c9f6ab24968401e450812fc0ff7d05b5ae07a63/pyproject.toml#L189-L202 for example configuration.

(from Joss review: https://github.com/openjournals/joss-reviews/issues/9316#issuecomment-3530069168)
