### 2025-11-17 unitaryfoundation/qlass#148 @FarLab
Automatic building and uploading releases to PyPi.
There is no CD workflow for automatically building and uploading releases to PyPI. There is some great tooling for this. See https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-sdist.yml, https://github.com/munich-quantum-toolkit/workflows/blob/main/.github/workflows/reusable-python-packaging-wheel-build.yml, and https://github.com/munich-quantum-toolkit/bench/blob/main/.github/workflows/cd.yml for example and workflows that you could be using.

(from Joss review: https://github.com/openjournals/joss-reviews/issues/9316#issuecomment-3530069168)
