### 2025-11-17 f2ae448
Add CD workflow for automatic PyPI deployment
