### 2025-11-17 unitaryfoundation/qlass#150 @goerz
Add citations to documentation
The documentation for this package would greatly benefit from citing some academic literature (and potentially also linking out to other material to provide background an explanations. There are a lot of very technical aspects to this package that would not be familiar even to an expert in quantum technology, so references would be important for clarification. I would probably  recommend using the [spinxcontrib-bibtex](https://sphinxcontrib-bibtex.readthedocs.io/en/latest/index.html) plugin for this.

Ref. https://github.com/openjournals/joss-reviews/issues/9316
