### 2025-11-17 unitaryfoundation/qlass#151 @goerz
Add CONTRIBUTING.md
In addition to the short "Contributing" section in the README. I would recommend to also have a longer `CONTRIBUTING.md` document. More important than pointers to various support channels or general notes on "how to use GitHub" would be a documentation of the development workflows used in this package. Anything that might be relevant for a new contributor/maintainer joining the project, or even a new external contributor looking to prepare a PR. How the code, tests and docs are organized, which tools are used for development, how to run the tests or build the documentation locally, how to format code, etc.. I had to look at your `ci.yml` to see that you use `pytest` to run tests, for example.

Ref https://github.com/openjournals/joss-reviews/issues/9316
