### 2025-11-17 unitaryfoundation/qlass#152 @goerz
Verify example code in CI
Example code like the snippets in the README, the files in the `notebooks` folder, and anything similar should be automatically verified as part of continuous integration (the pytest test suite). If runtimes are prohibitive, these verifications could also be kept separate so that they are automatic, albeit triggered on demand.

In any case, _how_ the examples are verified / should be kept up-to-date should be part of the `CONTRIBUTING.md` file (#151) and would also be something I'd need to understand in the context of https://github.com/openjournals/joss-reviews/issues/9316
