### 2025-11-18 a09d273
Update .github/workflows/cd.yml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
