+++ .github/workflows/cd.yml
@@ -52,4 +52,4 @@ jobs:
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
-          packages-dir: dist
+          packages_dir: dist
