+++ .github/workflows/cd.yml
@@ -41,6 +41,7 @@ jobs:
     runs-on: ubuntu-latest
     permissions:
       contents: read
+      id-token: write
     steps:
       - name: Download built distributions
         uses: actions/download-artifact@v4
