### 2025-11-18 a3e021a
update cd.yml to use best practices
