+++ .github/workflows/cd.yml
@@ -6,51 +6,54 @@ on:
   workflow_dispatch:
 
 env:
-  PYTHON_VERSION: "3.11"
+  PYTHON_VERSION: "3.12"
 
 jobs:
   build:
     runs-on: ubuntu-latest
-
+    
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
-
+      
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
           python-version: ${{ env.PYTHON_VERSION }}
           cache: pip
-
+      
       - name: Install build backend
         run: |
           python -m pip install --upgrade pip
           python -m pip install build
-
+      
       - name: Build distributions
         run: python -m build
-
+      
       - name: Upload distributions artifact
         uses: actions/upload-artifact@v4
         with:
           name: dist
           path: dist/*
+          retention-days: 5  # Clean up artifacts after 5 days
 
   publish:
     needs: build
     runs-on: ubuntu-latest
+    environment:
+      name: pypi  # Use GitHub environments for better security
+      url: https://pypi.org/p/${{ github.event.repository.name }}
     permissions:
       contents: read
-      id-token: write
+      id-token: write  # Required for trusted publishing (OIDC)
+    
     steps:
       - name: Download built distributions
         uses: actions/download-artifact@v4
         with:
           name: dist
           path: dist
-
+      
       - name: Publish to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          password: ${{ secrets.PYPI_API_TOKEN }}
-          packages_dir: dist
+        # No 'with' block needed when using trusted publishing
