### 2025-11-18 bde880a
add type checking using mypy in ci.yml
