+++ .github/workflows/ci.yml
@@ -12,6 +12,19 @@ permissions:
   id-token: write # Required for the `codecov/codecov-action` action
 
 jobs:
+  type-check:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Check out repository
+      uses: actions/checkout@v5
+
+    - name: Install the latest version of uv
+      uses: astral-sh/setup-uv@v7
+
+    - name: Run type checking
+      run: |
+        uv run --group type-check -p 3.12 mypy .
+
   test-and-coverage:
     runs-on: ubuntu-latest
     steps:
@@ -23,7 +36,7 @@ jobs:
 
     - name: Run tests and collect coverage
       run: |
-        uv run -p 3.12 pytest --cov --cov-report=xml
+        uv run --group test -p 3.12 pytest --cov --cov-report=xml
       env:
         # This disables Numba's JIT for accurate coverage reporting
         QLASS_DISABLE_JIT: 1
