### 2025-11-18 88960e2
fix type checking error in compiler.py
