+++ mypy.ini
@@ -1,0 +1,13 @@
+[mypy]
+# Global settings
+python_version = 3.12
+warn_return_any = True
+warn_unused_configs = True
+disallow_untyped_defs = True
+
+# Ignore missing stubs for these libraries
+[mypy-perceval.*]
+ignore_missing_imports = True
+
+[mypy-qiskit.*]
+ignore_missing_imports = True
