### 2025-11-18 f15d7c3
fix type check errors in quantum_chemistrym module
