+++ mypy.ini
@@ -11,3 +11,15 @@ ignore_missing_imports = True
 
 [mypy-qiskit.*]
 ignore_missing_imports = True
+
+[mypy-numba.*]
+ignore_missing_imports = True
+
+[mypy-openfermion.*]
+ignore_missing_imports = True
+
+[mypy-openfermionpyscf.*]
+ignore_missing_imports = True
+
+[mypy-pyscf.*]
+ignore_missing_imports = True
