### 2025-11-18 f5e49e8
fix type checking errors in utils.py
