+++ mypy.ini
@@ -23,3 +23,6 @@ ignore_missing_imports = True
 
 [mypy-pyscf.*]
 ignore_missing_imports = True
+
+[mypy-exqalibur.*]
+ignore_missing_imports = True
