### 2025-11-18 9047880
fix type checking errors in vqe module
