+++ mypy.ini
@@ -26,3 +26,9 @@ ignore_missing_imports = True
 
 [mypy-exqalibur.*]
 ignore_missing_imports = True
+
+[mypy-scipy.*]
+ignore_missing_imports = True
+
+[mypy-matplotlib.*]
+ignore_missing_imports = True
