### 2025-11-18 e53e435
exclude tests and example files for mypy
