+++ mypy.ini
@@ -5,6 +5,12 @@ warn_return_any = True
 warn_unused_configs = True
 disallow_untyped_defs = True
 
+# Exclude tests and examples from type checking
+exclude = (?x)(
+    ^tests/
+    | ^examples/
+)
+
 # Ignore missing stubs for these libraries
 [mypy-perceval.*]
 ignore_missing_imports = True
