### 2025-11-18 141b462
change ci.yml to only type check src/qlass
