+++ .github/workflows/ci.yml
@@ -9,7 +9,6 @@ on:
 
 permissions:
   contents: read
-  id-token: write # Required for the `codecov/codecov-action` action
 
 jobs:
   type-check:
@@ -43,8 +42,9 @@ jobs:
 
     - name: Upload coverage to Codecov
       if: github.event_name == 'pull_request'
-      uses: codecov/codecov-action@v5
+      uses: codecov/codecov-action@v4
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         files: ./coverage.xml
-        fail_ci_if_error: true
+        fail_ci_if_error: false
+        verbose: true
