### 2025-11-19 d22c046
add linter to ci
