+++ tests/test_vqe.py
@@ -1,22 +1,13 @@
+import warnings
 
 import numpy as np
 import perceval as pcvl
-from perceval.algorithm import Sampler
 import pytest
+from perceval.algorithm import Sampler
 
-from qlass.vqe import (
-    VQE,
-    le_ansatz,
-    custom_unitary_ansatz,
-    hf_ansatz
-)
+from qlass.quantum_chemistry import Hchain_KS_hamiltonian, LiH_hamiltonian
+from qlass.vqe import VQE, custom_unitary_ansatz, hf_ansatz, le_ansatz
 
-from qlass.quantum_chemistry import (
-    LiH_hamiltonian,
-    Hchain_KS_hamiltonian
-)
-
-import warnings
 warnings.simplefilter('ignore')
 warnings.filterwarnings('ignore')
 
@@ -138,9 +129,6 @@ def test_evqe_pipeline():
 
         return samples
 
-    # Number of qubits
-    num_qubits = 2
-
     ham, scf_mo_energy, n_orbs = Hchain_KS_hamiltonian(4, 1.2)
 
     vqe = VQE(
@@ -172,9 +160,6 @@ def test_invalid_cost_type():
 
         return samples
 
-    # Number of qubits
-    num_qubits = 2
-
     ham, scf_mo_energy, n_orbs = Hchain_KS_hamiltonian(4, 1.2)
 
     vqe = VQE(
@@ -184,7 +169,7 @@ def test_invalid_cost_type():
     )
 
     with pytest.raises(ValueError, match="Invalid cost option. Use 'VQE' or 'e-VQE'."):
-        vqe_energy = vqe.run(
+        vqe.run(
             max_iterations=5,
             verbose=True,
             weight_option="weighted",
@@ -196,9 +181,6 @@ def test_invalid_evqe_executor_type():
     def unitary_exec():
         return np.eye(4, dtype=complex)
 
-    # Number of qubits
-    num_qubits = 2
-
     ham, scf_mo_energy, n_orbs = Hchain_KS_hamiltonian(4, 1.2)
 
     vqe = VQE(
@@ -209,7 +191,7 @@ def test_invalid_evqe_executor_type():
     )
 
     with pytest.raises(ValueError, match="option: e-VQE takes only executor_type: sampling"):
-        vqe_energy = vqe.run(
+        vqe.run(
             max_iterations=5,
             verbose=True,
             weight_option="weighted",
@@ -240,7 +222,6 @@ def test_custom_unitary_ansatz():
 
     # Assert both logical outcomes are present and roughly balanced
     assert len(prob_dict) == 2, f"Unexpected number of outcomes: {prob_dict}"
-    keys = list(prob_dict.keys())
     assert all(state in prob_dict for state in [pcvl.BasicState('|1,0>'), pcvl.BasicState('|0,1>')]), \
         f"Expected states |1,0> and |0,1> not found in results: {prob_dict}"
 
@@ -352,7 +333,7 @@ def test_vqe_init_with_unitary_executor():
         num_params=2,
         executor_type="qubit_unitary"
     )
-    
+
     assert vqe.executor_type == "qubit_unitary"
     assert vqe.num_qubits == 2
     assert vqe.num_params == 2
@@ -441,31 +422,31 @@ def photonic_parametrized_executor(params):
     # Create a simple 4×4 unitary using rotation-like structure
     theta = params[0]
     phi = params[1] if len(params) > 1 else 0
-    
+
     # Simple parameterized unitary for 2 qubits (4 modes)
     U = np.eye(4, dtype=complex)
-    
+
     # Apply rotation-like transformation
     c, s = np.cos(theta), np.sin(theta)
[…]
