+++ src/qlass/__init__.py
@@ -1,23 +1,23 @@
 # export from compiler module
-from .compiler.compiler import compile, ResourceAwareCompiler, generate_report
+from .compiler.compiler import ResourceAwareCompiler, compile, generate_report
 from .compiler.hardware_config import HardwareConfig
 
 # export from problems module
-from .quantum_chemistry.classical_solution import hamiltonian_matrix, brute_force_minimize
+from .quantum_chemistry.classical_solution import brute_force_minimize, hamiltonian_matrix
 from .quantum_chemistry.hamiltonians import (
-    LiH_hamiltonian,
     Hchain_KS_hamiltonian,
-    transformation_Hmatrix_Hqubit,
-    pauli_commute, 
+    LiH_hamiltonian,
     group_commuting_pauli_terms,
-    sparsepauliop_dictionary
+    pauli_commute,
+    sparsepauliop_dictionary,
+    transformation_Hmatrix_Hqubit,
 )
 
-# export from vqe module
-from .vqe.ansatz import le_ansatz, custom_unitary_ansatz, hf_ansatz
+# export from utils module
+from .utils.utils import e_vqe_loss_function, loss_function, rotate_qubits
 
-# export from utils module 
-from .utils.utils import loss_function, rotate_qubits, e_vqe_loss_function
+# export from vqe module
+from .vqe.ansatz import custom_unitary_ansatz, hf_ansatz, le_ansatz
 
 # Define the public API exposed directly under 'qlass'
 __all__ = [
@@ -43,4 +43,5 @@ __all__ = [
 
 # Version information
 from importlib.metadata import version
+
 __version__ = version("qlass")

[…]
