+++ examples/vqe_example_pu_ps.py
@@ -12,11 +12,12 @@ from qlass.quantum_chemistry import LiH_hamiltonian_tapered, brute_force_minimiz
 from qlass.utils import linear_circuit_to_unitary
 from qlass.vqe import VQE
 
-warnings.simplefilter('ignore')
-warnings.filterwarnings('ignore')
+warnings.simplefilter("ignore")
+warnings.filterwarnings("ignore")
 
 # Initialize the Qiskit converter
 qiskit_converter = QiskitConverter(backend_name="Naive", noise_model=None)
+
 
 def executor(params):
     """
@@ -24,7 +25,7 @@ def executor(params):
     """
     num_qubits = 4
     # 1. Create the variational ansatz circuit
-    ansatz = n_local(num_qubits, 'ry', 'cx', reps=1, entanglement='linear')
+    ansatz = n_local(num_qubits, "ry", "cx", reps=1, entanglement="linear")
     ansatz_assigned = ansatz.assign_parameters(params)
 
     # Convert the circuit to a Perceval processor
@@ -33,6 +34,7 @@ def executor(params):
     unitary = linear_circuit_to_unitary(linear)
 
     return unitary
+
 
 def main():
     # Number of qubits
@@ -54,17 +56,16 @@ def main():
     vqe = VQE(
         hamiltonian=hamiltonian,
         executor=executor,
-        num_params=2*num_qubits, # Number of parameters in the linear entangled ansatz
+        num_params=2 * num_qubits,  # Number of parameters in the linear entangled ansatz
         executor_type="photonic_unitary",
-        ancillary_modes=list(range(8, 8+6)), # modes 0,1,..., 7 are for the qubits. need 6 more modes for 3 CNOTS.
+        ancillary_modes=list(
+            range(8, 8 + 6)
+        ),  # modes 0,1,..., 7 are for the qubits. need 6 more modes for 3 CNOTS.
     )
 
     # Run the VQE optimization
     print("\nRunning VQE optimization...")
-    vqe_energy = vqe.run(
-        max_iterations=100,
-        verbose=True
-    )
+    vqe_energy = vqe.run(max_iterations=100, verbose=True)
 
     # Get the optimal parameters
     optimal_params = vqe.get_optimal_parameters()
@@ -84,15 +85,16 @@ def main():
     print("\nPlotting convergence history...")
     plt.figure(figsize=(10, 6))
     iterations = range(len(vqe.energy_history))
-    plt.plot(iterations, vqe.energy_history, 'o-', label='VQE Energy')
-    plt.axhline(y=exact_energy, color='r', linestyle='--', label='Exact Energy')
-    plt.xlabel('Iteration')
-    plt.ylabel('Energy')
-    plt.title('VQE Convergence')
+    plt.plot(iterations, vqe.energy_history, "o-", label="VQE Energy")
+    plt.axhline(y=exact_energy, color="r", linestyle="--", label="Exact Energy")
+    plt.xlabel("Iteration")
+    plt.ylabel("Energy")
+    plt.title("VQE Convergence")
     plt.legend()
     plt.grid(True, alpha=0.3)
-    plt.savefig('vqe_convergence.png')
+    plt.savefig("vqe_convergence.png")
     print("Convergence plot saved as 'vqe_convergence.png'")
+
 
 if __name__ == "__main__":
     main()
