+++ tests/test_vqe.py
@@ -8,8 +8,8 @@ from perceval.algorithm import Sampler
 from qlass.quantum_chemistry import Hchain_KS_hamiltonian, LiH_hamiltonian
 from qlass.vqe import VQE, custom_unitary_ansatz, hf_ansatz, le_ansatz
 
-warnings.simplefilter('ignore')
-warnings.filterwarnings('ignore')
+warnings.simplefilter("ignore")
+warnings.filterwarnings("ignore")
 
 
 def mock_executor(params, pauli_string):
@@ -18,7 +18,8 @@ def mock_executor(params, pauli_string):
     For this test, the actual computation doesn't matter, only the format.
     It returns a distribution where '00' and '11' are equally likely.
     """
-    return {'counts': {'00': 500, '11': 500}}
+    return {"counts": {"00": 500, "11": 500}}
+
 
 @pytest.fixture
 def simple_vqe():
@@ -27,6 +28,7 @@ def simple_vqe():
     num_params = 4
     return VQE(hamiltonian=hamiltonian, executor=mock_executor, num_params=num_params)
 
+
 def test_vqe_init(simple_vqe):
     """Tests if the VQE class is initialized with the correct attributes."""
     assert simple_vqe.num_qubits == 2
@@ -34,6 +36,7 @@ def test_vqe_init(simple_vqe):
     assert simple_vqe.optimizer == "COBYLA"
     assert callable(simple_vqe.executor)
     assert simple_vqe.optimization_result is None
+
 
 def test_vqe_run(simple_vqe):
     """Tests the main `run` method to ensure optimization completes."""
@@ -47,6 +50,7 @@ def test_vqe_run(simple_vqe):
     assert len(simple_vqe.parameter_history) > 0
     # The final energy should match the one in the result object
     assert np.isclose(final_energy, simple_vqe.optimization_result.fun)
+
 
 def test_get_optimal_parameters(simple_vqe):
     """
@@ -65,6 +69,7 @@ def test_get_optimal_parameters(simple_vqe):
     assert len(optimal_params) == simple_vqe.num_params
     assert np.allclose(optimal_params, simple_vqe.optimization_result.x)
 
+
 def test_compare_with_exact(simple_vqe):
     """
     Tests the comparison with an exact energy value.
@@ -82,13 +87,13 @@ def test_compare_with_exact(simple_vqe):
     expected_abs_error = abs(vqe_energy - exact_energy)
     expected_rel_error = expected_abs_error / abs(exact_energy)
 
-    assert comparison['vqe_energy'] == vqe_energy
-    assert comparison['exact_energy'] == exact_energy
-    assert np.isclose(comparison['absolute_error'], expected_abs_error)
-    assert np.isclose(comparison['relative_error'], expected_rel_error)
+    assert comparison["vqe_energy"] == vqe_energy
+    assert comparison["exact_energy"] == exact_energy
+    assert np.isclose(comparison["absolute_error"], expected_abs_error)
+    assert np.isclose(comparison["relative_error"], expected_rel_error)
+
 
 def test_vqe_pipeline():
-
     # Define an executor function that uses the linear entangled ansatz
     def executor(params, pauli_string):
         processor = le_ansatz(params, pauli_string)
@@ -106,23 +111,19 @@ def test_vqe_pipeline():
     vqe = VQE(
         hamiltonian=hamiltonian,
         executor=executor,
-        num_params=2*num_qubits, # Number of parameters in the linear entangled ansatz
+        num_params=2 * num_qubits,  # Number of parameters in the linear entangled ansatz
     )
 
     # Run the VQE optimization
-    vqe_energy = vqe.run(
[…]
