+++ src/qlass/utils/__init__.py
@@ -1,4 +1,3 @@
-
 from .utils import (
     compute_energy,
     compute_expectation_value_from_unitary,
@@ -33,5 +32,4 @@ __all__ = [
     "logical_state_to_modes",
     "photon_to_qubit_unitary",
     "loss_function_photonic_unitary",
-
 ]
