### 2025-11-24 1aff189
Update perceval-quandela requirement
Updates the requirements on [perceval-quandela](https://github.com/Quandela/Perceval) to permit the latest version.
- [Release notes](https://github.com/Quandela/Perceval/releases)
- [Commits](https://github.com/Quandela/Perceval/compare/v0.13.2...v1.1.0)

---
updated-dependencies:
- dependency-name: perceval-quandela
  dependency-version: 1.1.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
