+++ .github/workflows/ci.yml
@@ -59,7 +59,7 @@ jobs:
 
     - name: Upload coverage to Codecov
       if: github.event_name == 'pull_request'
-      uses: codecov/codecov-action@v4
+      uses: codecov/codecov-action@v5
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         files: ./coverage.xml
