+++ .github/workflows/cd.yml
@@ -49,7 +49,7 @@ jobs:
     
     steps:
       - name: Download built distributions
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v6
         with:
           name: dist
           path: dist
