### 2025-11-24 5350dcd
Update hamiltonian for H2 derived from WFT
