+++ tests/test_quantum_chemistry.py
@@ -17,8 +17,10 @@ from typing import Dict
 import pytest
 from openfermion import QubitOperator
 import warnings
+
 warnings.simplefilter('ignore')
 warnings.filterwarnings('ignore')
+
 
 def test_pauli_string_to_matrix():
     """
@@ -40,22 +42,24 @@ def test_pauli_string_to_matrix():
     assert np.allclose(pauli_string_to_matrix("ZY"), np.kron(Z, Y))
     assert np.allclose(pauli_string_to_matrix("XX"), np.kron(X, X))
 
+
 def test_hamiltonian_matrix():
     """
     Tests the conversion of a Hamiltonian dictionary to its matrix representation.
     """
     X = np.array([[0, 1], [1, 0]], dtype=complex)
     Z = np.array([[1, 0], [0, -1]], dtype=complex)
-    
+
     hamiltonian_dict = {"X": 0.5, "Z": -0.5}
-    
+
     # Expected matrix: 0.5 * X + (-0.5) * Z
     expected_matrix = 0.5 * X - 0.5 * Z
-    
+
     result_matrix = hamiltonian_matrix(hamiltonian_dict)
-    
+
     assert np.allclose(result_matrix, expected_matrix)
     assert result_matrix.shape == (2, 2)
+
 
 def test_brute_force_minimize():
     """
@@ -75,8 +79,9 @@ def test_brute_force_minimize():
     }
     # Using numpy to get the exact value for comparison
     exact_min_eig = np.min(np.linalg.eigvalsh(hamiltonian_matrix(hamiltonian_2q)))
-    
+
     assert np.isclose(brute_force_minimize(hamiltonian_2q), exact_min_eig)
+
 
 def test_eig_decomp_lanczos():
     """
@@ -88,21 +93,22 @@ def test_eig_decomp_lanczos():
     np.random.seed(42)
     random_real_matrix = np.random.rand(dim, dim)
     hermitian_matrix = (random_real_matrix + random_real_matrix.T.conj()) / 2
-    
+
     # Convert the matrix to complex type to match the vector type in Lanczos
     hermitian_matrix_complex = hermitian_matrix.astype(np.complex128)
 
     # Get exact eigenvalues from NumPy
     exact_eigenvalues = np.linalg.eigvalsh(hermitian_matrix_complex)
-    
+
     # Get eigenvalues from Lanczos implementation
     lanczos_eigenvalues = eig_decomp_lanczos(hermitian_matrix_complex)
-    
+
     # Sort both sets of eigenvalues for comparison
     exact_eigenvalues.sort()
     lanczos_eigenvalues.sort()
-    
+
     assert np.allclose(exact_eigenvalues[0], lanczos_eigenvalues[0], atol=1e-3)
+
 
 def test_lanczos_tridiagonalization():
     """
@@ -139,6 +145,7 @@ def test_lanczos_tridiagonalization():
     rhs = V.T @ T
     assert np.allclose(lhs, rhs), "Tridiagonal matrix T does not satisfy the Lanczos relation"
 
+
 def test_lanczos_early_exit():
     """
     Tests the early exit condition of the lanczos function when beta becomes zero.
@@ -151,7 +158,7 @@ def test_lanczos_early_exit():
     # An initial vector that is an eigenvector of A.
     # The Krylov subspace will be 1-dimensional, forcing beta to be zero on the second iteration.
     v_init = np.array([0, 0, 1, 0], dtype=np.complex128)
-    
+
     # We expect the loop to break after the first iteration (j=1)
     m = dim
     T, V = lanczos(A, v_init, m=m)
@@ -160,19 +167,20 @@ def test_lanczos_early_exit():
[…]
