### 2025-11-24 c6cdb82
Updating WFT Hamiltonian for Hchain
