+++ src/qlass/quantum_chemistry/hamiltonians.py
@@ -1,15 +1,21 @@
 import itertools
+
 import numpy as np
+
 # OpenFermion imports - replacing qiskit_nature
 from openfermion.chem import MolecularData
-from openfermion.transforms import get_fermion_operator, jordan_wigner, symmetry_conserving_bravyi_kitaev
+
 # Import QubitOperator para type hinting
 from openfermion.ops import InteractionOperator, QubitOperator
+from openfermion.transforms import (
+    get_fermion_operator,
+    jordan_wigner,
+    symmetry_conserving_bravyi_kitaev,
+)
 from openfermionpyscf import run_pyscf
 
-from typing import Dict, List, Tuple
 
-def sparsepauliop_dictionary(H: QubitOperator) -> Dict[str, float]:
+def sparsepauliop_dictionary(H: QubitOperator) -> dict[str, float]:
     """
     Converts an OpenFermion QubitOperator into a dictionary representation.
 
@@ -27,13 +33,13 @@ def sparsepauliop_dictionary(H: QubitOperator) -> Dict[str, float]:
         representing a term in the Hamiltonian, and its value is the
         real part of the coefficient for that term.
     """
-    pauli_dict: Dict[str, float] = {}
+    pauli_dict: dict[str, float] = {}
 
     # Determine the total number of qubits in the system.
     # This is found by identifying the highest qubit index acted upon by any Pauli operator.
-    max_qubit_idx = -1 # Initialize to -1 to correctly handle 0-indexed qubits
+    max_qubit_idx = -1  # Initialize to -1 to correctly handle 0-indexed qubits
     if H.terms:
-        for pauli_term_key in H.terms.keys():
+        for pauli_term_key in H.terms:
             if pauli_term_key:  # Checks if the term is not the global identity `()`
                 # pauli_term_key is a tuple of (qubit_index, Pauli_operator_char) tuples, e.g., ((0, 'X'), (1, 'Y'))
                 current_max_for_term = max(idx for idx, _ in pauli_term_key)
@@ -41,45 +47,48 @@ def sparsepauliop_dictionary(H: QubitOperator) -> Dict[str, float]:
                     max_qubit_idx = current_max_for_term
         num_qubits = max_qubit_idx + 1
     else:
-        num_qubits = max_qubit_idx + 1 # If max_qubit_idx remains -1, num_qubits becomes 0.
+        num_qubits = max_qubit_idx + 1  # If max_qubit_idx remains -1, num_qubits becomes 0.
 
     # Let's ensure num_qubits is at least 1 if an identity term is present and no other terms define size.
-    if not H.terms and num_qubits == 0 : # Truly empty QubitOperator
-        return {} # An empty operator has no Pauli terms.
-    if H.terms and not any(bool(term) for term in H.terms.keys()) and num_qubits == 0:
+    if not H.terms and num_qubits == 0:  # Truly empty QubitOperator
+        return {}  # An empty operator has no Pauli terms.
+    if H.terms and not any(bool(term) for term in H.terms) and num_qubits == 0:
         # This means H.terms only contains {(): coeff}, e.g. QubitOperator('')
         # max_qubit_idx was -1, num_qubits became 0. For an identity string, we need at least 1 qubit.
         num_qubits = 1
-        
+
     # A term consists of a Pauli product (pauli_string_openfermion) and its coefficient.
     for pauli_string_openfermion, coefficient in H.terms.items():
-        if not pauli_string_openfermion:  # This is the global identity term, represented by an empty tuple `()`.
+        if (
+            not pauli_string_openfermion
+        ):  # This is the global identity term, represented by an empty tuple `()`.
             # The Pauli key for the identity is a string of 'I's, one for each qubit.
-            pauli_key = 'I' * num_qubits
+            pauli_key = "I" * num_qubits
[…]
