+++ tests/test_quantum_chemistry.py
@@ -1,44 +1,45 @@
-from qlass.quantum_chemistry import (
-    LiH_hamiltonian,
-    generate_random_hamiltonian,
-    LiH_hamiltonian_tapered,
-    pauli_string_to_matrix,
-    hamiltonian_matrix,
-    Hchain_KS_hamiltonian,
-    transformation_Hmatrix_Hqubit,
-    brute_force_minimize,
-    eig_decomp_lanczos,
-    lanczos,
-    group_commuting_pauli_terms_openfermion_hybrid,
-    pauli_commute
-)
-import numpy as np
-from typing import Dict
-import pytest
-from openfermion import QubitOperator
 import warnings
 
-warnings.simplefilter('ignore')
-warnings.filterwarnings('ignore')
+import numpy as np
+import pytest
+from openfermion import QubitOperator
+
+from qlass.quantum_chemistry import (
+    Hchain_KS_hamiltonian,
+    LiH_hamiltonian,
+    LiH_hamiltonian_tapered,
+    brute_force_minimize,
+    eig_decomp_lanczos,
+    generate_random_hamiltonian,
+    group_commuting_pauli_terms_openfermion_hybrid,
+    hamiltonian_matrix,
+    lanczos,
+    pauli_commute,
+    pauli_string_to_matrix,
+    transformation_Hmatrix_Hqubit,
+)
+
+warnings.simplefilter("ignore")
+warnings.filterwarnings("ignore")
 
 
 def test_pauli_string_to_matrix():
     """
     Tests the conversion of Pauli strings to their matrix representations.
     """
-    I = np.eye(2, dtype=complex)
+    I_pauli = np.eye(2, dtype=complex)
     X = np.array([[0, 1], [1, 0]], dtype=complex)
     Y = np.array([[0, -1j], [1j, 0]], dtype=complex)
     Z = np.array([[1, 0], [0, -1]], dtype=complex)
 
     # Test single-qubit operators
-    assert np.allclose(pauli_string_to_matrix("I"), I)
+    assert np.allclose(pauli_string_to_matrix("I"), I_pauli)
     assert np.allclose(pauli_string_to_matrix("X"), X)
     assert np.allclose(pauli_string_to_matrix("Y"), Y)
     assert np.allclose(pauli_string_to_matrix("Z"), Z)
 
     # Test two-qubit operators (tensor product)
-    assert np.allclose(pauli_string_to_matrix("IX"), np.kron(I, X))
+    assert np.allclose(pauli_string_to_matrix("IX"), np.kron(I_pauli, X))
     assert np.allclose(pauli_string_to_matrix("ZY"), np.kron(Z, Y))
     assert np.allclose(pauli_string_to_matrix("XX"), np.kron(X, X))
 
@@ -73,10 +74,7 @@ def test_brute_force_minimize():
     # Eigenvalues of XX are +1, +1, -1, -1
     # Eigenvalues of IZ are +1, -1, +1, -1
     # This Hamiltonian's ground state energy is known to be approx -0.5385
-    hamiltonian_2q = {
-        "XX": 0.5,
-        "IZ": 0.2
-    }
+    hamiltonian_2q = {"XX": 0.5, "IZ": 0.2}
     # Using numpy to get the exact value for comparison
     exact_min_eig = np.min(np.linalg.eigvalsh(hamiltonian_matrix(hamiltonian_2q)))
 
@@ -119,12 +117,9 @@ def test_lanczos_tridiagonalization():
     """
     # 1. Set up a sample Hermitian matrix and initial vector
     dim = 4
-    A = np.array([
-        [2, -1, 0, 0],
-        [-1, 2, -1, 0],
[…]
