### 2025-11-24 629c219
Update hamiltonians.py
