### 2025-11-24 da4143a
Update test_quantum_chemistry.py
