### 2025-11-24 cfc38a0
Update ensemble_vqe_H2.py
