### 2025-11-24 3959b8d
Update hamiltonians.py
