### 2025-11-24 9f7d3a1
Update ensemble_vqe_H2.py
