+++ examples/ensemble_vqe_H2.py
@@ -1,10 +1,11 @@
+import warnings
 import numpy as np
+import matplotlib.pyplot as plt
 from qlass.quantum_chemistry.hamiltonians import Hchain_hamiltonian_WFT, hamiltonian_matrix
 from qlass.vqe import VQE
-from perceval.algorithm import Sampler
-import warnings
-import matplotlib.pyplot as plt
 from qlass.vqe.ansatz import hf_ansatz
+from perceval.algorithm import Sampler
+warnings.filterwarnings('ignore')
 
 warnings.filterwarnings('ignore')
 
