### 2025-11-24 a1112bd
Update ensemble_vqe_H2.py
