### 2025-11-24 eaa0c30
Update ensemble_vqe_H2.py
