### 2025-11-24 6a1a2d3
Update ensemble_vqe_H2.py
