### 2025-11-24 8c247cc
Update ensemble_vqe_H2.py
