### 2025-11-24 4e07c2d
Update ensemble_vqe_H2.py
