+++ examples/ensemble_vqe_H2.py
@@ -1,10 +1,11 @@
-# Standard library
+
 import warnings
-# Third-party packages
+
 import matplotlib.pyplot as plt
 import numpy as np
-# Local or project-specific imports
+
 from perceval.algorithm import Sampler
+
 from qlass.quantum_chemistry.hamiltonians import Hchain_hamiltonian_WFT, hamiltonian_matrix
 from qlass.vqe import VQE
 from qlass.vqe.ansatz import hf_ansatz
