### 2025-11-24 083bd73
Update ensemble_vqe_H2.py
