### 2025-11-24 e7c8916
Update ensemble_vqe_H2.py
