+++ examples/ensemble_vqe_H2.py
@@ -1,4 +1,3 @@
-
 import warnings
 
 import matplotlib.pyplot as plt
