+++ src/qlass/quantum_chemistry/hamiltonians.py
@@ -522,69 +522,75 @@ def transformation_Hmatrix_Hqubit(Hmatrix: np.ndarray, nqubits: int) -> QubitOpe
     return H_qubit
 
 
-def Hchain_hamiltonian_WFT(n_hydrogens: int = 2, R: float = 0.8, charge: int = 0, spin: int = 0, num_electrons: int = 2, num_orbitals: int = 2) -> dict[str, float]:
+def Hchain_hamiltonian_WFT(
+    n_hydrogens: int = 2,
+    R: float = 0.8,
+    charge: int = 0,
+    spin: int = 0,
+    num_electrons: int = 2,
+    num_orbitals: int = 2,
+) -> dict[str, float]:
     """
-Construct the qubit Hamiltonian for a linear hydrogen chain (Hₙ) using
-    wavefunction-based methods.
+    Construct the qubit Hamiltonian for a linear hydrogen chain (Hₙ) using
+        wavefunction-based methods.
 
-    This function builds the molecular geometry, performs a PySCF electronic
-    structure calculation through OpenFermion, extracts an active-space molecular
-    Hamiltonian, and maps it to a qubit Hamiltonian.
+        This function builds the molecular geometry, performs a PySCF electronic
+        structure calculation through OpenFermion, extracts an active-space molecular
+        Hamiltonian, and maps it to a qubit Hamiltonian.
 
-    Nuclear repulsion energy is included manually. The resulting Hamiltonian is
-    returned as a dictionary mapping Pauli strings to coefficients.
+        Nuclear repulsion energy is included manually. The resulting Hamiltonian is
+        returned as a dictionary mapping Pauli strings to coefficients.
 
-    Parameters
-    ----------
-    n_hydrogens : int, optional
-        Number of hydrogen atoms in the linear chain. Must be even, as atoms
-        are paired symmetrically about the origin. Default is ``2``.
-    R : float, optional
-        Bond length between adjacent hydrogens in ångström. Default is ``0.8``.
-    charge : int, optional
-        Total molecular charge. Default is ``0``.
-    spin : int, optional
-        Spin multiplicity parameter such that multiplicity = ``2S + 1``.
-        For example, ``spin=0`` corresponds to a singlet. Default is ``0``.
-    num_electrons : int, optional
-        Number of electrons.
-        Default is ``2``.
-    num_orbitals : int, optional
-        Number of spatial molecular orbitals.
-        Default is ``2``.
+        Parameters
+        ----------
+        n_hydrogens : int, optional
+            Number of hydrogen atoms in the linear chain. Must be even, as atoms
+            are paired symmetrically about the origin. Default is ``2``.
+        R : float, optional
+            Bond length between adjacent hydrogens in ångström. Default is ``0.8``.
+        charge : int, optional
+            Total molecular charge. Default is ``0``.
+        spin : int, optional
+            Spin multiplicity parameter such that multiplicity = ``2S + 1``.
+            For example, ``spin=0`` corresponds to a singlet. Default is ``0``.
+        num_electrons : int, optional
+            Number of electrons.
+            Default is ``2``.
+        num_orbitals : int, optional
+            Number of spatial molecular orbitals.
+            Default is ``2``.
 
-    Returns
-    -------
-    Dict[str, float]
-        A dictionary representing the qubit Hamiltonian, where keys are
-        Pauli strings (e.g., ``"XIZY"``) and values are real coefficients.
+        Returns
+        -------
+        Dict[str, float]
+            A dictionary representing the qubit Hamiltonian, where keys are
[…]
