### 2025-11-24 08bbd8c
add relevant references in the documentation
