+++ docs/refs.bib
@@ -1,0 +1,53 @@
+@article{peruzzo2014variational,
+  title={A variational eigenvalue solver on a photonic quantum processor},
+  author={Peruzzo, Alberto and McClean, Jarrod and Shadbolt, Peter and Yung, Man-Hong and Zhou, Xiao-Qi and Love, Peter J and Aspuru-Guzik, Al{\'a}n and O’brien, Jeremy L},
+  journal={Nature communications},
+  volume={5},
+  number={1},
+  pages={4213},
+  year={2014},
+  publisher={Nature Publishing Group}
+}
+
+@article{heurtel2023perceval,
+  title={Perceval: A software platform for discrete variable photonic quantum computing},
+  author={Heurtel, Nicolas and Fyrillas, Andreas and de Gliniasty, Gr{\'e}goire and Le Bihan, Rapha{\"e}l and Malherbe, S{\'e}bastien and Prawer, Marceau and Solal, Edouard and Valiron, Beno{\^\i}t and Yeh, Peter and Mansfield, Shane},
+  journal={Quantum},
+  volume={7},
+  pages={931},
+  year={2023},
+  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
+}
+
+@article{kok2007linear,
+  title={Linear optical quantum computing with photonic qubits},
+  author={Kok, Pieter and Munro, William J and Nemoto, Kae and Ralph, Timothy C and Dowling, Jonathan P and Milburn, Gerard J},
+  journal={Reviews of Modern Physics},
+  volume={79},
+  number={1},
+  pages={135},
+  year={2007},
+  publisher={APS}
+}
+
+@article{kohn1965self,
+  title={Self-consistent equations including exchange and correlation effects},
+  author={Kohn, Walter and Sham, Lu Jeu},
+  journal={Physical review},
+  volume={140},
+  number={4A},
+  pages={A1133},
+  year={1965},
+  publisher={APS}
+}
+
+@article{nakanishi2019subspace,
+  title={Subspace-search variational quantum eigensolver for excited states},
+  author={Nakanishi, Ken M and Mitarai, Kosuke and Fujii, Keisuke},
+  journal={Physical Review Research},
+  volume={1},
+  number={3},
+  pages={033062},
+  year={2019},
+  publisher={APS}
+}
