### 2025-11-24 0007ec9
update pyproject.toml and fix formatting
