### 2025-11-24 #9316 @danielskatz
👋 @zimboras - please use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
