### 2025-11-28 60e0292
ci: add workflow to run example scripts
