### 2025-11-28 13cdacc
fix import in ensemble_vqe_H2.py
