### 2025-11-28 030366f
reduce run time of example scripts
