+++ examples/vqe_unitary_executor_example.py
@@ -66,7 +66,7 @@ def main():
     print(f"Relative Error: {comparison['relative_error']:.2%}")
 
     # Plot convergence
-    vqe.plot_convergence(exact_energy=exact_energy)
+    # vqe.plot_convergence(exact_energy=exact_energy)
 
 
 if __name__ == "__main__":
