+++ examples/vqe_simulation_qiskit.py
@@ -84,17 +84,17 @@ def main():
     print(f"Energy difference: {comparison['absolute_error']:.6f}")
 
     # Plot convergence
-    plt.figure(figsize=(10, 6))
-    plt.plot(
-        range(len(vqe.energy_history)), vqe.energy_history, "o-", label="VQE Energy (Qiskit Sim)"
-    )
-    plt.axhline(y=exact_energy, color="r", linestyle="--", label="Exact Energy")
-    plt.xlabel("Iteration")
-    plt.ylabel("Energy")
-    plt.title("VQE Convergence using Qiskit Executor")
-    plt.grid(True, alpha=0.3)
-    plt.legend()
-    plt.show()
+    # plt.figure(figsize=(10, 6))
+    # plt.plot(
+    #     range(len(vqe.energy_history)), vqe.energy_history, "o-", label="VQE Energy (Qiskit Sim)"
+    # )
+    # plt.axhline(y=exact_energy, color="r", linestyle="--", label="Exact Energy")
+    # plt.xlabel("Iteration")
+    # plt.ylabel("Energy")
+    # plt.title("VQE Convergence using Qiskit Executor")
+    # plt.grid(True, alpha=0.3)
+    # plt.legend()
+    # plt.show()
 
 
 if __name__ == "__main__":
