+++ examples/vqe_example.py
@@ -67,18 +67,18 @@ def main():
     print(f"Energy difference: {comparison['absolute_error']:.6f}")
 
     # Plot the convergence
-    print("\nPlotting convergence history...")
-    plt.figure(figsize=(10, 6))
-    iterations = range(len(vqe.energy_history))
-    plt.plot(iterations, vqe.energy_history, "o-", label="VQE Energy")
-    plt.axhline(y=exact_energy, color="r", linestyle="--", label="Exact Energy")
-    plt.xlabel("Iteration")
-    plt.ylabel("Energy")
-    plt.title("VQE Convergence")
-    plt.legend()
-    plt.grid(True, alpha=0.3)
-    plt.savefig("vqe_convergence.png")
-    print("Convergence plot saved as 'vqe_convergence.png'")
+    # print("\nPlotting convergence history...")
+    # plt.figure(figsize=(10, 6))
+    # iterations = range(len(vqe.energy_history))
+    # plt.plot(iterations, vqe.energy_history, "o-", label="VQE Energy")
+    # plt.axhline(y=exact_energy, color="r", linestyle="--", label="Exact Energy")
+    # plt.xlabel("Iteration")
+    # plt.ylabel("Energy")
+    # plt.title("VQE Convergence")
+    # plt.legend()
+    # plt.grid(True, alpha=0.3)
+    # plt.savefig("vqe_convergence.png")
+    # print("Convergence plot saved as 'vqe_convergence.png'")
 
 
 if __name__ == "__main__":

[…]
