+++ examples/ensemble_vqe_H2.py
@@ -1,6 +1,5 @@
 import warnings
 
-import matplotlib.pyplot as plt
 import numpy as np
 from perceval.algorithm import Sampler
 
