### 2025-11-28 unitaryfoundation/qlass#152 @goerz
You’d want to make sure that there’s a script or documented command that verifies all the examples. That doesn’t need to run on every CI push automatically. It’s fine as something you can run manually on demand. Personally, I like having a Makefile (`make check-examples`) for this kind of thing, but that can be a bit UNIX-centric. So your mileage may vary.
