### 2025-12-01 4075c11
remove examples.yml workflow file
