+++ .github/workflows/examples.yml
@@ -1,33 +1,0 @@
-name: Run example scripts
-
-on:
-  pull_request:
-    paths:
-      - 'examples/*.py'
-      - '.github/workflows/examples.yml'
-  workflow_dispatch:
-
-jobs:
-  run-examples:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.11]
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install .
-      - name: Run all example scripts
-        run: |
-          set -e
-          for script in examples/*.py; do
-            echo "Running $script"
-            python "$script"
-          done
