### 2025-12-01 0f5b76a
added a Makefile for running all example files as a check
