+++ examples/ensemble_vqe_H2.py
@@ -39,6 +39,8 @@ vqe_energy = vqe.run(max_iterations=20, verbose=True, weight_option="weighted", 
 H_matrix = hamiltonian_matrix(ham)
 exact_energy = np.sort(np.linalg.eigvals(H_matrix))
 print(f"energy from exact diag: {exact_energy}")
+
+# Uncomment the following block to show a plot of the result
 # plt.figure(figsize=(10, 6))
 
 # plt.plot(vqe.energy_collector.loss_data, label="cost")
