+++ examples/ensemble_vqe_H4.py
@@ -38,6 +38,7 @@ H_matrix = hamiltonian_matrix(ham)
 exact_energy = np.sort(np.linalg.eigvals(H_matrix))
 
 # Uncomment the following block to show a plot of the result
+# import matplotlib.pyplot as plt
 # plt.figure(figsize=(10, 6))
 
 # plt.plot(vqe.energy_collector.loss_data, label="Cost")
