### 2025-12-01 cde6c9e
Update examples/ensemble_vqe_H2.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
