+++ examples/ensemble_vqe_H2.py
@@ -50,7 +50,7 @@ print(f"energy from exact diag: {exact_energy}")
 # plt.axhline(y=exact_energy[0], color="b", linestyle="--", label="E_0")
 # plt.axhline(y=exact_energy[1], color="g", linestyle="--", label="E_1")
 # plt.xlabel("Iteration")
-# plt.ylabel("Energy (Hartee)")
+# plt.ylabel("Energy (Hartree)")
 # plt.title("ensemble VQE Convergence")
 # plt.legend()
 # plt.tight_layout()
