### 2025-12-01 a4fbf57
Update examples/noisy_example.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
