### 2025-12-03 0b444cd
Update installation instructions to use uv instead of pip
