### 2025-12-04 131674b
Update CONTRIBUTING.md
Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
