### 2025-12-04 48f9e83
Update CONTRIBUTING.md
Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
