### 2025-12-04 bff209b
Update CONTRIBUTING.md
Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
