### 2025-12-04 99d9812
Update CONTRIBUTING.md
Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
