### 2025-12-04 6d226ac
Update CONTRIBUTING.md
Co-authored-by: Nathan Shammah <nathan.shammah@gmail.com>
