### 2025-12-05 5476026
Update installation.rst
