### 2025-12-05 d94986f
Update installation.rst
