### 2025-12-08 252dc68
hf_ansatz seperated as Bitstring_initial_states and CSF_initial_states to avoid any confusion
