+++ src/qlass/vqe/ansatz.py
@@ -108,33 +108,156 @@ def list_of_ones(computational_basis_state: int, n_qubits: int) -> list[int]:
     return [abs(j - n_qubits + 1) for j in range(len(bitstring)) if bitstring[j] == "1"]
 
 
-def hf_ansatz(
-    layers: int,
-    n_orbs: int,
-    lp: np.ndarray,
+def CSF_initial_states(
+    num_spatial_orbitals: int,
+    num_electrons: [int, int],
+    initial_parameters: np.ndarray,
     pauli_string: str,
-    method: str,
-    cost: str = "VQE",
+    singlet_excitation: bool = False,
+    k: int | None = None,
+    l: int | None = None,
     noise_model: NoiseModel | None = None,
 ) -> Processor | list[Processor]:
     """
-    Build a Hartree–Fock-based variational ansatz using Qiskit's ``n_local`` circuit,
+    Generate a Hartree-Fock initial state quantum circuit and optionally apply a singlet excitation.
+
+    This function constructs a Hartree-Fock initial state for a given number of electrons
+    and spatial orbitals, applies a parameterized ansatz, and prepares the circuit for
+    simulation on a quantum processor. Optionally, it can include a singlet excitation
+    between specified orbitals.
+
+    Parameters
+    ----------
+    num_spatial_orbitals : int
+        Number of spatial orbitals in the system.
+    num_electrons : list of int
+        Number of alpha and beta electrons, given as [num_alpha, num_beta].
+    initial_parameters : np.ndarray
+        Initial values for the parameterized ansatz.
+    pauli_string : str
+        Pauli string used to rotate the qubits after the ansatz.
+    singlet_excitation : bool, optional
+        If True, apply a singlet excitation to the Hartree-Fock state using the specified
+        orbitals `i` and `j`. Default is False.
+    i : int or None, optional
+        Index of the occupied orbital for singlet excitation. Required if `singlet_excitation=True`.
+    j : int or None, optional
+        Index of the unoccupied orbital for singlet excitation. Required if `singlet_excitation=True`.
+    noise_model : NoiseModel or None, optional
+        Optional noise model for simulating the quantum processor. Default is None.
+
+    Returns
+    -------
+    Processor or list of Processor
+        - If `singlet_excitation=False`, returns a single `Processor` object representing the
+          Hartree-Fock initial state with the ansatz applied.
+        - If `singlet_excitation=True`, returns a list `[Processor_HF, Processor_SC]` containing
+          the Hartree-Fock processor and the singlet-excited processor.
+
+    Raises
+    ------
+    ValueError
+        If `singlet_excitation=True` but either `k` or `l` is not provided.
+
+    Notes
+    -----
+    - The Hartree-Fock state is created by initializing qubits corresponding to spin orbitals.
+    - Do not use tampered Hamiltonian on this function.
+    - The ansatz used is a `n_local` circuit with 'ry' rotations and linear 'cx' entanglement.
+    - The circuit is transpiled using basis gates ['u3', 'cx'] and optimized to level 3.
+    - The `rotate_qubits` function is applied to align with the specified Pauli string.
+    - This function integrates with Perceval's `LogicalState` and `compile` functions to
+      produce a processor-ready quantum circuit.
+    """
+    # Generating Hartree Fock initial state
+    num_spin_orbitals = sum(num_electrons) * 2
+    bitstring_alpha = [0] * (num_spin_orbitals // 2)
+    bitstring_beta = [0] * (num_spin_orbitals // 2)
+    for i in range(num_electrons[0]):
[…]
