+++ examples/ensemble_vqe_H4.py
@@ -1,11 +1,12 @@
 import warnings
+
 import matplotlib.pyplot as plt
 import numpy as np
 
 from perceval.algorithm import Sampler
 
-from qlass.vqe import VQE
 from qlass.quantum_chemistry import Hchain_KS_hamiltonian, hamiltonian_matrix
+from qlass.vqe import VQE
 from qlass.vqe.ansatz import Bitstring_initial_states
 
 warnings.filterwarnings("ignore")
