+++ examples/ensemble_vqe_H2.py
@@ -1,12 +1,13 @@
 import warnings
+
 import matplotlib.pyplot as plt
 import numpy as np
 
 from perceval.algorithm import Sampler
 
+from qlass.quantum_chemistry import hamiltonian_matrix
 from qlass.quantum_chemistry.hamiltonians import Hchain_hamiltonian_WFT
 from qlass.vqe import VQE
-from qlass.quantum_chemistry import hamiltonian_matrix
 from qlass.vqe.ansatz import Bitstring_initial_states
 
 
