### 2025-12-08 ca076cc
ruff formatting
