+++ examples/vqe_example_HF_ansatz.py
@@ -41,7 +41,7 @@ plt.figure(figsize=(10, 6))
 plt.plot(vqe.energy_history, label="Ground state from VQE")
 plt.axhline(y=exact_energy[0], color="b", linestyle="--", label="Ground state from exact")
 plt.xlabel("Iteration")
-plt.ylabel("Energy (Hartee)")
+plt.ylabel("Energy (Hartree)")
 plt.title("VQE Convergence")
 plt.legend()
 plt.tight_layout()
