+++ examples/ensemble_vqe_H4.py
@@ -53,7 +53,7 @@ ax[1].plot(error_EA, color="b", ls=":", label="E_A")
 ax[1].plot(error_EB, color="g", ls=":", label="E_B")
 ax[1].set_yscale("log")
 ax[0].set_ylabel("Energy (hartree)", fontsize=16)
-ax[1].set_ylabel("Error (hartree", fontsize=16)
+ax[1].set_ylabel("Error (hartree)", fontsize=16)
 ax[1].set_xlabel("Iteration", fontsize=16)
 ax[0].tick_params(axis="x", which="both", bottom=True, top=False, labelbottom=False)
 ax[1].tick_params(axis="both", labelsize=16)
