### 2025-12-10 083d3af
Update vqe.py
Verbose: printing cost and energies while using "e-VQE"
