### 2025-12-16 unitaryfund/qlass#173 @goerz
Statement of Need in the Documentation
https://github.com/openjournals/joss-reviews/issues/9316 asks that the software documentation also includes the "Statement of Need". Personally, I don't think it needs (of even should have) that title (it reads a bit unnatural). But I do think having the kind of information contained in the "Statement of Need" in the paper, specifically a brief survey and comparison with other related software is a good thing to have in the documentation.
