### 2025-12-16 3c193b3
add some statement of need for the package in the introduction of the documentation
