+++ docs/introduction.rst
@@ -1,7 +1,7 @@
 Introduction
 ============
 
-Photonic quantum computing represents a powerful approach to building scalable quantum computers, offering advantages such as avoiding the need for dilution fridges, providing natural connectivity to quantum communication networks [@kok2007linear; @o2009photonic] and, as in the case of QLASS devices, allowing for reconfigurable chips. However, the translation of quantum algorithms from the abstract circuit model to photonic implementations presents unique challenges. Existing quantum software frameworks like Qiskit :cite:p:`aleksandrowicz2019qiskit` and Cirq :cite:p:`cirq2021` primarily target gate-based quantum computers with qubit architectures, while photonic platforms operate on fundamentally different principles using linear optical elements, photons, and modes. 
+Photonic quantum computing represents a powerful approach to building scalable quantum computers, offering advantages such as avoiding the need for dilution fridges, providing natural connectivity to quantum communication networks :cite:p:`kok2007linear, o2009photonic` and, as in the case of QLASS devices, allowing for reconfigurable chips. However, the translation of quantum algorithms from the abstract circuit model to photonic implementations presents unique challenges. Existing quantum software frameworks like Qiskit :cite:p:`aleksandrowicz2019qiskit` and Cirq :cite:p:`cirq2021` primarily target gate-based quantum computers with qubit architectures, while photonic platforms operate on fundamentally different principles using linear optical elements, photons, and modes. 
 
 There are only a few open source software projects that focus on the simulation or compilation of photonic quantum computers. These include Perceval :cite:p:`heurtel2023perceval`, Strawberry  Fields :cite:p:`killoran2019strawberry`, Piquasso :cite:p:`kolarovszki2025piquasso` and Graphix :cite:p:`sunami2022graphix`, all Python packages. These tools are primarily focussed on photonic circuit compilation and simulation. We are however interested in specific applications of photonic quantum computing. In particular in applications related to quantum chemistry. As far as we know, there is no software platform that can run VQE simulations on photonic quantum computing platforms starting with an ansatz defined in the qubit architecture all the way to running the VQE on a photonic quantum computer to obtain ground state energies for molecular Hamiltonians.
 
