### 2025-12-19 #9316 @FarLab
@goerz I added a paragraph in the introduction of the documentation expressing a statement of need. Let me know if that is what you had in mind!
