+++ docs/conf.py
@@ -10,7 +10,7 @@
 project = "qlass"
 copyright = "2025, Farrokh Labib, Nathan Shammah"
 author = "Farrokh Labib, Nathan Shammah"
-release = "0.2.0"
+release = "0.2.1"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
