### 2026-01-05 #9316 @goerz
> @goerz - does this satisfy your request?

Yes, I think the paper is fine to publish at this point. I've check off all items on my list
