### 2026-01-12 69b55e9
minor changes in paper.md
