+++ paper/paper.md
@@ -32,13 +32,13 @@ bibliography: paper.bib
 
 Photonic quantum computing represents a powerful approach to building scalable quantum computers, offering advantages such as avoiding the need for dilution fridges, providing natural connectivity to quantum communication networks [@kok2007linear; @o2009photonic] and, as in the case of QLASS devices, allowing for reconfigurable chips. However, the translation of quantum algorithms from the abstract circuit model to photonic implementations presents unique challenges. Existing quantum software frameworks like Qiskit [@aleksandrowicz2019qiskit] and Cirq [@cirq2021] primarily target gate-based quantum computers with qubit architectures, while photonic platforms operate on fundamentally different principles using linear optical elements, photons, and modes. 
 
-There are only a few open source software projects that focus on the simulation or compilation of photonic quantum computers. These include Perceval [@heurtel2023perceval], Strawberry  Fields [@killoran2019strawberry], Piquasso [@kolarovszki2025piquasso] and Graphix [@sunami2022graphix], all Python packages. These tools are primarily focussed on photonic circuit compilation and simulation. We are however interested in specific applications of photonic quantum computing. In particular in applications related to quantum chemistry. As far as we know, there is no software platform that can run VQE simulations on photonic quantum computing platforms starting with an ansatz defined in the qubit architecture all the way to running the VQE on a photonic quantum computer to obtain ground state energies for molecular Hamiltonians.
+There are only a few open source software projects that focus on the simulation or compilation of photonic quantum computers. These include Perceval [@heurtel2023perceval], Strawberry  Fields [@killoran2019strawberry], Piquasso [@kolarovszki2025piquasso], and Graphix [@sunami2022graphix], all Python packages. These tools are primarily focussed on photonic circuit compilation and simulation. We are, however, interested in specific applications of photonic quantum computing, in particular, in applications related to quantum chemistry. As far as we know, there is no software platform that can run VQE simulations on photonic quantum computing platforms starting with an ansatz defined in the qubit architecture all the way to running the VQE on a photonic quantum computer to obtain ground state energies for molecular Hamiltonians.
 
 `qlass` addresses this gap by providing:
 
-1. **Quantum chemistry integration** with established packages (OpenFermion, PySCF) for molecular problem definition.
-2. **Optimized VQE implementations** specifically designed for the constraints and capabilities of photonic quantum computers.
-3. **Hardware-specific resource estimation and compilation** from quantum circuits (Qiskit) to linear optics layout of photonic processors (Perceval), including the analysis of circuits against realistic hardware constraints such as photon loss, detector efficiency, and fusion gate success rates.
+1. **Quantum chemistry integration** with established packages (OpenFermion, PySCF) for molecular problem definition
+2. **Optimized VQE implementations** specifically designed for the constraints and capabilities of photonic quantum computers
+3. **Hardware-specific resource estimation and compilation** from quantum circuits (Qiskit) to linear optics layout of photonic processors (Perceval), including the analysis of circuits against realistic hardware constraints such as photon loss, detector efficiency, and fusion gate success rates
 
 The package is particularly valuable for researchers working at the intersection of quantum algorithms and photonic quantum computing, as it provides the necessary tools to evaluate algorithm performance under realistic hardware conditions and optimize implementations for photonic architectures.
 
@@ -68,32 +68,33 @@ Key features include resource-aware compilation with realistic hardware modeling
 
 The package is structured into four main modules:
 
-1. **`compiler`**: Handles circuit translation and resource analysis.
-2. **`quantum_chemistry`**: Provides Hamiltonian generation and classical solutions.
-3. **`vqe`**: Implements variational algorithms and ansätze.
-4. **`utils`**: Contains utility functions for measurement processing and expectation value calculations.
[…]
