+++ paper/paper.bib
@@ -59,7 +59,7 @@
 @book{cirq2021,
   title={Cirq}, 
   url={https://zenodo.org/doi/10.5281/zenodo.4062499}, 
-  DOI={10.5281/ZENODO.4062499}, 
+  DOI={10.5281/zenodo.4062499}, 
   note={Python package for writing, manipulating, and running quantum circuits on quantum computers and simulators.}, 
   publisher={Zenodo}, 
   author={{Cirq Developers}}, 
