### 2026-01-13 8c4568f
Update paper.bib
