+++ paper/paper.bib
@@ -46,7 +46,7 @@
 
 @article{o2009photonic,
   title={Photonic quantum technologies},
-  author={O'brien, Jeremy L and Furusawa, Akira and Vu{\v{c}}kovi{\'c}, Jelena},
+  author={O'Brien, Jeremy L and Furusawa, Akira and Vu{\v{c}}kovi{\'c}, Jelena},
   journal={Nature photonics},
   volume={3},
   number={12},
