### 2026-01-14 3b49861
Update paper.md
