### 2026-01-14 #9316 @nathanshammah
Thank you @danielskatz, here is the Zenodo reference [10.5281/zenodo.18245793](https://doi.org/10.5281/zenodo.18245793). The tag is v0.2.2, https://github.com/unitaryfoundation/qlass/releases/tag/v0.2.2.
Can you please check the submission section is fine in the paper draft as we did not edit it?
