### 2026-01-14 #9316 @danielskatz
@nathanshammah & @FarLab

Once you take care of this 👇 , I can accept the paper.

> [@nathanshammah](https://github.com/nathanshammah)
>
> > Please check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
> Note that you can edit the metadata without changing the DOI / making a new deposit
